### 2022-09-30 #4810 @editorialbot
[PRE REVIEW]: Diart: A Python Library for Real-Time Speaker Diarization
**Submitting author:** @juanmc2005 (<a href="http://orcid.org/0000-0002-5035-147X">Juan Manuel Coria</a>)
**Repository:** https://github.com/juanmc2005/StreamingSpeakerDiarization
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.5.1
