### 2022-10-05 a2de3ab
Add real time inference and benchmark hooks
