+++ src/diart/inference.py
@@ -1,7 +1,7 @@
 import logging
 from pathlib import Path
 from traceback import print_exc
-from typing import Union, Text, Optional, Callable, Tuple, List
+from typing import Union, Text, Optional, Any, Tuple, List, Sequence
 
 import diart.operators as dops
 import diart.sources as src
@@ -10,13 +10,33 @@ import pandas as pd
 import rx
 import rx.operators as ops
 from diart.blocks import OnlineSpeakerDiarization, Resample
-from diart.sinks import DiarizationPredictionAccumulator, RTTMWriter, RealTimePlot, WindowClosedException
+from diart.sinks import DiarizationPredictionAccumulator, RealTimePlot, WindowClosedException
 from diart.utils import Chronometer
 from pyannote.core import Annotation, SlidingWindowFeature
 from pyannote.database.util import load_rttm
 from pyannote.metrics.diarization import DiarizationErrorRate
 from rx.core import Observer
 from tqdm import tqdm
+
+
+class RealTimeInferenceHook:
+    def on_audio_chunk(self, waveform: SlidingWindowFeature):
+        pass
+
+    def on_audio_batch(self, batch: Sequence[SlidingWindowFeature]):
+        pass
+
+    def on_prediction_batch(self, batch: Sequence[Tuple[Annotation, SlidingWindowFeature]]):
+        pass
+
+    def on_prediction(self, prediction: Tuple[Annotation, SlidingWindowFeature]):
+        pass
+
+    def on_completion(self, prediction: Annotation):
+        pass
+
+    def on_error(self, error: BaseException):
+        pass
 
 
 class RealTimeInference:
@@ -60,12 +80,15 @@ class RealTimeInference:
         show_progress: bool = True,
         progress_desc: Optional[Text] = None,
         leave_progress_bar: bool = False,
+        hooks: Optional[Sequence[RealTimeInferenceHook]] = None,
     ):
         self.pipeline = pipeline
         self.source = source
         self.batch_size = batch_size
         self.do_profile = do_profile
         self.do_plot = do_plot
+        # TODO replace profiling and progress bar with hooks
+        self.hooks = [] if hooks is None else hooks
         self.accumulator = DiarizationPredictionAccumulator(source.uri)
         self._chrono = Chronometer("chunk" if self.batch_size == 1 else "batch")
         self._observers = []
@@ -94,7 +117,9 @@ class RealTimeInference:
         # Add rx operators to manage the inputs and outputs of the pipeline
         self.stream = self.stream.pipe(
             dops.rearrange_audio_stream(chunk_duration, step_duration, sample_rate),
+            ops.do_action(lambda wav: self._call_hooks("on_audio_chunk", wav)),
             ops.buffer_with_count(count=self.batch_size),
+            ops.do_action(lambda batch: self._call_hooks("on_audio_batch", batch)),
         )
 
         if self.do_profile:
@@ -107,7 +132,9 @@ class RealTimeInference:
             self.stream = self.stream.pipe(ops.map(pipeline))
 
         self.stream = self.stream.pipe(
+            ops.do_action(lambda pred_batch: self._call_hooks("on_prediction_batch", pred_batch)),
             ops.flat_map(lambda results: rx.from_iterable(results)),
+            ops.do_action(lambda pred: self._call_hooks("on_prediction", pred)),
             ops.do(self.accumulator),
         )
 
@@ -120,6 +147,11 @@ class RealTimeInference:
                 ops.do_action(on_next=lambda _: self._pbar.update())
             )
 
+    def _call_hooks(self, method: Text, arg: Any):
+        """Call ``method`` on every attached hook with parameter ``arg``"""
+        for hook in self.hooks:
[…]
