### 2022-10-05 6417799
Clean some code in Benchmark
