### 2022-10-11 bf4738c
Add online diarization pipeline hooks
