### 2022-10-18 320d239
Add Ego4D report and hyper-parameters
