### 2022-10-18 b407549
added cropping_mode to DelayedAggregation
