+++ src/diart/blocks/aggregation.py
@@ -102,6 +102,8 @@ class DelayedAggregation:
         "mean": simple average
         "hamming": average weighted by the Hamming window values (aligned to the buffer)
         "any": no aggregation, pick the first overlapping window
+    cropping_mode: ("strict", "loose", "center"), optional
+        Defines the cropping mode. Defaults to "loose".
 
     Example
     --------
@@ -130,10 +132,12 @@ class DelayedAggregation:
         step: float,
         latency: Optional[float] = None,
         strategy: Literal["mean", "hamming", "first"] = "hamming",
+        cropping_mode: Literal["strict", "loose", "center"] = "loose"
     ):
         self.step = step
         self.latency = latency
         self.strategy = strategy
+        self.cropping_mode = cropping_mode
 
         if self.latency is None:
             self.latency = self.step
@@ -159,7 +163,7 @@ class DelayedAggregation:
             num_frames = output_window.data.shape[0]
             first_region = Segment(0, output_region.end)
             first_output = buffers[0].crop(
-                first_region, fixed=first_region.duration
+                first_region, mode=self.cropping_mode, fixed=first_region.duration
             )
             first_output[-num_frames:] = output_window.data
             resolution = output_region.end / first_output.shape[0]
