### 2022-10-19 a4f7352
added assert for cropping mode
Co-authored-by: Juan Coria <juanmc2005@hotmail.com>
