### 2022-10-19 e7d6dd7
added details about the cropping mode to the documentation of DelayedAggregation
