+++ src/diart/blocks/aggregation.py
@@ -104,6 +104,9 @@ class DelayedAggregation:
         "any": no aggregation, pick the first overlapping window
     cropping_mode: ("strict", "loose", "center"), optional
         Defines the cropping mode. Defaults to "loose".
+        This will define the value of the mode parameter used in SlidingWindowFeature.crop
+        (from pyannote.core, see https://pyannote.github.io/pyannote-core/reference.html#pyannote.core.SlidingWindowFeature.crop
+        for more details).
 
     Example
     --------
