### 2022-10-19 4866586
added cropping_mode to AggregationStrategy and children
