### 2022-10-19 6e94676
Add missing cropping_mode in aggregation calls
