### 2022-10-19 6e0f27a
Add missing cropping_mode in aggregation calls
