### 2022-10-19 81c2b96
Add new contributor to author list
