+++ src/diart/argdoc.py
@@ -11,3 +11,4 @@ MAX_SPEAKERS = "Maximum number of speakers"
 CPU = "Force models to run on CPU"
 BATCH_SIZE = "For segmentation and embedding pre-calculation. If BATCH_SIZE < 2, run fully online and estimate real-time latency"
 OUTPUT = "Directory to store the system's output in RTTM format"
+HF_TOKEN = "Huggingface authentication token for hosted models ('true' | 'false' | <token>). If 'true', it will use the token from huggingface-cli login"
