+++ setup.cfg
@@ -28,7 +28,9 @@ install_requires=
     einops>=0.3.0
     tqdm>=4.64.0
     pandas>=1.4.2
+    torch>=1.12.1
     torchaudio>=0.12.1,<1.0
+    pyannote.audio>=2.1.1
     pyannote.core>=4.5
     pyannote.database>=4.1.1
     pyannote.metrics>=3.2
