+++ src/diart/tune.py
@@ -1,14 +1,14 @@
 import argparse
 from pathlib import Path
 
+import diart.argdoc as argdoc
 import optuna
 import torch
-from optuna.samplers import TPESampler
-
-import diart.argdoc as argdoc
+from diart import utils
+from diart.blocks import PipelineConfig
 from diart.models import SegmentationModel, EmbeddingModel
 from diart.optim import Optimizer, HyperParameter
-from diart.blocks import PipelineConfig
+from optuna.samplers import TPESampler
 
 
 def run():
@@ -37,10 +37,15 @@ def run():
     parser.add_argument("--storage", type=str,
                         help="Optuna storage string. If provided, continue a previous study instead of creating one. The database name must match the study name")
     parser.add_argument("--output", type=str, help="Working directory")
+    parser.add_argument("--hf-token", default="true", type=str,
+                        help=f"{argdoc.HF_TOKEN}. Defaults to 'true' (required by pyannote)")
     args = parser.parse_args()
     args.device = torch.device("cpu") if args.cpu else None
-    args.segmentation = SegmentationModel.from_pyannote(args.segmentation)
-    args.embedding = EmbeddingModel.from_pyannote(args.embedding)
+    args.hf_token = utils.parse_hf_token_arg(args.hf_token)
+
+    # Download pyannote models (or get from cache)
+    args.segmentation = SegmentationModel.from_pyannote(args.segmentation, args.hf_token)
+    args.embedding = EmbeddingModel.from_pyannote(args.embedding, args.hf_token)
 
     # Create the base configuration for each trial
     base_config = PipelineConfig.from_namespace(args)
