### 2022-10-31 f8fdace
Fix README. Change version to 0.6.0
