+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name=diart
-version=0.5.1
+version=0.6.0
 author=Juan Manuel Coria
 description=Speaker diarization in real time
 long_description=file: README.md
