+++ README.md
@@ -232,7 +232,7 @@ mic = MicrophoneAudioSource(sample_rate)
 
 stream = mic.stream.pipe(
     # Reformat stream to 5s duration and 500ms shift
-    dops.rearrange_audio_stream(sample_rate),
+    dops.rearrange_audio_stream(sample_rate=sample_rate),
     ops.map(lambda wav: (wav, segmentation(wav))),
     ops.starmap(embedding)
 ).subscribe(on_next=lambda emb: print(emb.shape))
