### 2022-10-31 e81917c
Simplify installation instructions in README.md
