### 2022-11-15 79bf2f2
Add parameters to some hook methods
