+++ src/diart/blocks/diarization.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from typing import Optional, Any, Union, Tuple, Sequence
 
 import numpy as np
@@ -8,10 +10,10 @@ from typing_extensions import Literal
 from .aggregation import DelayedAggregation
 from .clustering import OnlineSpeakerClustering
 from .embedding import OverlapAwareSpeakerEmbedding
-from ..features import TemporalFeatures
 from .segmentation import SpeakerSegmentation
 from .utils import Binarize
 from .. import models as m
+from ..features import TemporalFeatures
 
 
 class PipelineConfig:
@@ -89,6 +91,7 @@ class PipelineConfig:
 class OnlineSpeakerDiarizationHook:
     def on_local_segmentation_batch(
         self,
+        pipeline: OnlineSpeakerDiarization,
         audio_batch: torch.Tensor,
         segmentation_batch: TemporalFeatures
     ):
@@ -96,6 +99,7 @@ class OnlineSpeakerDiarizationHook:
 
     def on_embedding_batch(
         self,
+        pipeline: OnlineSpeakerDiarization,
         audio_batch: torch.Tensor,
         embedding_batch: torch.Tensor
     ):
@@ -103,39 +107,56 @@ class OnlineSpeakerDiarizationHook:
 
     def on_local_segmentation(
         self,
+        pipeline: OnlineSpeakerDiarization,
         waveform: SlidingWindowFeature,
         segmentation: SlidingWindowFeature
     ):
         pass
 
-    def on_embeddings(self, waveform: SlidingWindowFeature, embeddings: torch.Tensor):
+    def on_embeddings(
+        self,
+        pipeline: OnlineSpeakerDiarization,
+        waveform: SlidingWindowFeature,
+        embeddings: torch.Tensor
+    ):
         pass
 
     def on_before_clustering(
         self,
-        waveform: SlidingWindowFeature,
-        clustering: OnlineSpeakerClustering
+        pipeline: OnlineSpeakerDiarization,
+        waveform: SlidingWindowFeature
     ):
         pass
 
     def on_after_clustering(
         self,
+        pipeline: OnlineSpeakerDiarization,
         waveform: SlidingWindowFeature,
         clustering: OnlineSpeakerClustering,
         segmentation: SlidingWindowFeature
     ):
         pass
 
-    def on_soft_prediction(self, waveform: SlidingWindowFeature, segmentation: SlidingWindowFeature):
+    def on_soft_prediction(
+        self,
+        pipeline: OnlineSpeakerDiarization,
+        waveform: SlidingWindowFeature,
+        segmentation: SlidingWindowFeature
+    ):
         pass
 
-    def on_binary_prediction(self, waveform: SlidingWindowFeature, diarization: Annotation):
+    def on_binary_prediction(
+        self,
+        pipeline: OnlineSpeakerDiarization,
+        waveform: SlidingWindowFeature,
+        diarization: Annotation
+    ):
         pass
 
-    def on_before_reset(self):
+    def on_before_reset(self, pipeline: OnlineSpeakerDiarization):
         pass
 
-    def on_after_reset(self):
+    def on_after_reset(self, pipeline: OnlineSpeakerDiarization,):
         pass
 
 
@@ -170,13 +191,13 @@ class OnlineSpeakerDiarization:
         self.binarize = Binarize(self.config.tau_active)
 
         # Internal state, handle with care
-        self.clustering = None
+        self.clustering: Optional[OnlineSpeakerClustering] = None
         self.chunk_buffer, self.pred_buffer = [], []
         self.reset()
 
     def reset(self):
         for hook in self.hooks:
-            hook.on_before_reset()
+            hook.on_before_reset(self)
 
         self.clustering = OnlineSpeakerClustering(
             self.config.tau_active,
[…]
