### 2023-01-31 ed0ebb3
Fix some minor things that could cause problems in the future
