+++ src/diart/inference.py
@@ -20,8 +20,7 @@ from tqdm import tqdm
 
 
 class RealTimeInference:
-    """
-    Performs inference in real time given a pipeline and an audio source.
+    """Performs inference in real time given a pipeline and an audio source.
     Streams an audio source to an online speaker diarization pipeline.
     It allows users to attach a chain of operations in the form of hooks.
 
@@ -66,7 +65,7 @@ class RealTimeInference:
         self.batch_size = batch_size
         self.do_profile = do_profile
         self.do_plot = do_plot
-        self.accumulator = DiarizationPredictionAccumulator(source.uri)
+        self.accumulator = DiarizationPredictionAccumulator(self.source.uri)
         self._chrono = Chronometer("chunk" if self.batch_size == 1 else "batch")
         self._observers = []
 
@@ -76,19 +75,19 @@ class RealTimeInference:
 
         # Estimate the total number of chunks that the source will emit
         self.num_chunks = None
-        if source.duration is not None:
-            numerator = source.duration - chunk_duration + step_duration
+        if self.source.duration is not None:
+            numerator = self.source.duration - chunk_duration + step_duration
             self.num_chunks = int(np.ceil(numerator / step_duration))
 
         self.stream = self.source.stream
 
         # Dynamic resampling if the audio source isn't compatible
-        if sample_rate != source.sample_rate:
-            msg = f"Audio source has sample rate {source.sample_rate}, " \
+        if sample_rate != self.source.sample_rate:
+            msg = f"Audio source has sample rate {self.source.sample_rate}, " \
                   f"but pipeline's is {sample_rate}. Will resample."
             logging.warning(msg)
             self.stream = self.stream.pipe(
-                ops.map(Resample(source.sample_rate, sample_rate))
+                ops.map(Resample(self.source.sample_rate, sample_rate))
             )
 
         # Add rx operators to manage the inputs and outputs of the pipeline
@@ -100,11 +99,11 @@ class RealTimeInference:
         if self.do_profile:
             self.stream = self.stream.pipe(
                 ops.do_action(on_next=lambda _: self._chrono.start()),
-                ops.map(pipeline),
+                ops.map(self.pipeline),
                 ops.do_action(on_next=lambda _: self._chrono.stop()),
             )
         else:
-            self.stream = self.stream.pipe(ops.map(pipeline))
+            self.stream = self.stream.pipe(ops.map(self.pipeline))
 
         self.stream = self.stream.pipe(
             ops.flat_map(lambda results: rx.from_iterable(results)),
@@ -114,7 +113,7 @@ class RealTimeInference:
         # Show progress if required
         self._pbar = None
         if show_progress:
-            desc = f"Streaming {source.uri}" if progress_desc is None else progress_desc
+            desc = f"Streaming {self.source.uri}" if progress_desc is None else progress_desc
             self._pbar = tqdm(desc=desc, total=self.num_chunks, unit="chunk", leave=leave_progress_bar)
             self.stream = self.stream.pipe(
                 ops.do_action(on_next=lambda _: self._pbar.update())
