### 2023-03-01 #4810 @hadware
Yeah sure, but i've worked on pyannote (on which this project relies). This might also interest @hbredin
