### 2023-03-01 #4810 @juanmc2005
@Fei-Tao wouldn't that be a conflict of interest as well? @hadware has worked on pyannote.audio, which is owned by @hbredin (co-author)
