### 2023-03-06 #4810 @juanmc2005
Hi @sneakers-the-rat, you can focus on the `main` branch, whose last commit is [here](https://github.com/juanmc2005/StreamingSpeakerDiarization/commit/e81917c3adb6cd18adc99bda48401635122aa014)
