### 2023-03-08 397a386
Initial multithreading implementation with rich in Benchmark
