### 2023-03-08 14d9fed
Initial multithreading implementation with rich in Benchmark
