+++ src/diart/inference.py
@@ -1,4 +1,5 @@
 import logging
+import time
 from pathlib import Path
 from traceback import print_exc
 from typing import Union, Text, Optional, Callable, Tuple, List
@@ -11,12 +12,13 @@ import rx
 import rx.operators as ops
 from diart.blocks import BasePipeline, Resample
 from diart.sinks import DiarizationPredictionAccumulator, RealTimePlot, WindowClosedException
-from diart.utils import Chronometer
+from diart.utils import Chronometer, ProgressBar, ParallelProgressBars
 from pyannote.core import Annotation, SlidingWindowFeature
 from pyannote.database.util import load_rttm
 from pyannote.metrics.diarization import DiarizationErrorRate
 from rx.core import Observer
-from tqdm import tqdm
+from copy import deepcopy
+from concurrent.futures import ThreadPoolExecutor
 
 
 class RealTimeInference:
@@ -57,14 +59,14 @@ class RealTimeInference:
         do_profile: bool = True,
         do_plot: bool = False,
         show_progress: bool = True,
-        progress_desc: Optional[Text] = None,
-        leave_progress_bar: bool = False,
+        progress_bar: Optional[ProgressBar] = None,
     ):
         self.pipeline = pipeline
         self.source = source
         self.batch_size = batch_size
         self.do_profile = do_profile
         self.do_plot = do_plot
+        self.show_progress = show_progress
         self.accumulator = DiarizationPredictionAccumulator(self.source.uri)
         self._chrono = Chronometer("chunk" if self.batch_size == 1 else "batch")
         self._observers = []
@@ -78,6 +80,13 @@ class RealTimeInference:
         if self.source.duration is not None:
             numerator = self.source.duration - chunk_duration + step_duration
             self.num_chunks = int(np.ceil(numerator / step_duration))
+
+        # Show progress if required
+        self._pbar = progress_bar
+        if self.show_progress:
+            if self._pbar is None:
+                self._pbar = ProgressBar(default_description=f"Streaming {self.source.uri}")
+            self._pbar.create(total=self.num_chunks)
 
         self.stream = self.source.stream
 
@@ -110,11 +119,7 @@ class RealTimeInference:
             ops.do(self.accumulator),
         )
 
-        # Show progress if required
-        self._pbar = None
         if show_progress:
-            desc = f"Streaming {self.source.uri}" if progress_desc is None else progress_desc
-            self._pbar = tqdm(desc=desc, total=self.num_chunks, unit="chunk", leave=leave_progress_bar)
             self.stream = self.stream.pipe(
                 ops.do_action(on_next=lambda _: self._pbar.update())
             )
@@ -161,12 +166,12 @@ class RealTimeInference:
         interrupted = isinstance(error, KeyboardInterrupt)
         if not window_closed and not interrupted:
             print_exc()
-        # Close progress and chronometer states
+        # Close internal states
         self._close_pbar()
         self._close_chronometer()
 
     def _handle_completion(self):
-        # Close progress and chronometer states
+        # Close internal states
         self._close_pbar()
         self._close_chronometer()
 
@@ -178,6 +183,8 @@ class RealTimeInference:
         predictions: Annotation
             Speaker diarization pipeline predictions
         """
+        if self.show_progress:
+            self._pbar.start()
         config = self.pipeline.config
         observable = self.stream
         if self.do_plot:
@@ -241,6 +248,7 @@ class Benchmark:
         show_progress: bool = True,
         show_report: bool = True,
         batch_size: int = 32,
+        num_workers: int = 2,
     ):
         self.speech_path = Path(speech_path).expanduser()
         assert self.speech_path.is_dir(), "Speech path must be a directory"
@@ -262,6 +270,33 @@ class Benchmark:
[…]
