+++ requirements.txt
@@ -14,3 +14,4 @@ pyannote.database>=4.1.1
 pyannote.metrics>=3.2
 optuna>=2.10
 websockets>=10.3
+rich>=12.5.1
