### 2023-03-09 da570c9
Add multiprocessing as a 'Parallelize' wrapper. Add docs. Add example in README
