+++ src/diart/benchmark.py
@@ -3,8 +3,8 @@ from pathlib import Path
 
 import diart.argdoc as argdoc
 import pandas as pd
-from diart.blocks import OnlineSpeakerDiarization
-from diart.inference import Benchmark
+from diart.blocks import OnlineSpeakerDiarization, PipelineConfig
+from diart.inference import Benchmark, Parallelize
 
 
 def run():
@@ -41,10 +41,15 @@ def run():
         show_progress=True,
         show_report=True,
         batch_size=args.batch_size,
-        num_workers=args.num_workers,
     )
 
-    report = benchmark(OnlineSpeakerDiarization, vars(args))
+    if args.num_workers > 0:
+        benchmark = Parallelize(benchmark, args.num_workers)
+        report = benchmark(OnlineSpeakerDiarization, vars(args))
+    else:
+        config = PipelineConfig.from_dict(vars(args))
+        report = benchmark(OnlineSpeakerDiarization(config))
+
     if args.output is not None and isinstance(report, pd.DataFrame):
         report.to_csv(args.output / "benchmark_report.csv")
 
