### 2023-03-09 d801e6e
Add multiprocessing as a 'Parallelize' wrapper. Add docs. Add example in README
