+++ src/diart/inference.py
@@ -11,7 +11,7 @@ import pandas as pd
 import rx
 import rx.operators as ops
 import torch
-from diart.blocks import BasePipeline, BasePipelineConfig, Resample
+from diart.blocks import BasePipeline, Resample
 from diart.progress import ProgressBar, RichProgressBar, TQDMProgressBar
 from diart.sinks import DiarizationPredictionAccumulator, RealTimePlot, WindowClosedException
 from diart.utils import Chronometer
@@ -216,9 +216,9 @@ class RealTimeInference:
 class Benchmark:
     """
     Run an online speaker diarization pipeline on a set of audio files in batches.
-    It writes predictions to a given output directory.
+    Write predictions to a given output directory.
 
-    If the reference is given, it calculates the average diarization error rate.
+    If the reference is given, calculate the average diarization error rate.
 
     Parameters
     ----------
@@ -245,9 +245,6 @@ class Benchmark:
         embeddings, running the rest in real time.
         The performance between this two modes does not differ.
         Defaults to 32.
-    num_workers: int
-        Number of parallel workers.
-        Defaults to 0 (no parallelism).
     """
     def __init__(
         self,
@@ -257,7 +254,6 @@ class Benchmark:
         show_progress: bool = True,
         show_report: bool = True,
         batch_size: int = 32,
-        num_workers: int = 0,
     ):
         self.speech_path = Path(speech_path).expanduser()
         assert self.speech_path.is_dir(), "Speech path must be a directory"
@@ -279,18 +275,61 @@ class Benchmark:
         self.show_progress = show_progress
         self.show_report = show_report
         self.batch_size = batch_size
-        self.num_workers = num_workers
 
     @staticmethod
-    def _get_pbar_description(filepath: Path, idx_file: int, num_files: int) -> Text:
+    def get_pbar_description(filepath: Path, idx_file: int, num_files: int) -> Text:
+        """Return the description of the progress bar for the run on this file.
+
+        Parameters
+        ----------
+        filepath: Path
+            Path to the target file.
+        idx_file: int
+            The file number within the entire dataset.
+        num_files: int
+            The total number of files in the dataset.
+
+        Returns
+        -------
+        description: Text
+            Description for the progress bar.
+        """
         return f"Streaming {filepath.stem} ({idx_file + 1}/{num_files})"
 
-    def _run_inference(
+    def get_file_paths(self) -> List[Path]:
+        """Return the path for each file in the benchmark.
+
+        Returns
+        -------
+        paths: List[Path]
+            List of audio file paths.
+        """
+        return list(self.speech_path.iterdir())
+
+    def run_single(
         self,
         pipeline: BasePipeline,
         filepath: Path,
-        progress_bar: ProgressBar
-    ):
+        progress_bar: ProgressBar,
+    ) -> Annotation:
+        """Run a given pipeline on a given file.
+        Note that this method does NOT reset the
+        state of the pipeline before execution.
+
+        Parameters
+        ----------
+        pipeline: BasePipeline
+            Speaker diarization pipeline to run.
+        filepath: Path
+            Path to the target file.
+        progress_bar: diart.progress.ProgressBar
+            An object to manage the progress of this run.
+
+        Returns
+        -------
+        prediction: Annotation
+            Pipeline prediction for the given file.
+        """
         stream_padding = pipeline.config.latency - pipeline.config.step
         block_size = int(np.rint(pipeline.config.step * pipeline.config.sample_rate))
         source = src.FileAudioSource(filepath, pipeline.config.sample_rate, stream_padding, block_size)
@@ -313,30 +352,35 @@ class Benchmark:
 
         return pred
 
-    def _run_inference_job(
-        self,
-        pipeline_class: type,
-        config_dict: Dict[Text, Hashable],
-        filepath: Path,
[…]
