+++ setup.cfg
@@ -36,6 +36,7 @@ install_requires=
     pyannote.metrics>=3.2
     optuna>=2.10
     websockets>=10.3
+    rich>=12.5.1
 
 [options.packages.find]
 where=src
