### 2023-03-10 737c3ac
Send configuration object to Benchmark child processes instead of dictionary. Refactoring of diart.models
