### 2023-03-10 abe5f66
Send configuration object to Benchmark child processes instead of dictionary. Refactoring of diart.models
