+++ src/diart/models.py
@@ -4,16 +4,82 @@ import torch
 import torch.nn as nn
 
 try:
-    import pyannote.audio.pipelines.utils as pyannote
+    import pyannote.audio.pipelines.utils as pyannote_loader
     _has_pyannote = True
 except ImportError:
     _has_pyannote = False
+
+
+class LazyModel(nn.Module):
+    @property
+    def model(self) -> Optional[nn.Module]:
+        raise NotImplementedError
+
+    def load(self):
+        """Load model to memory"""
+        raise NotImplementedError
+
+    def is_in_memory(self) -> bool:
+        """Return whether the model has been loaded into memory"""
+        return self.model is not None
+
+    def to(self, *args, **kwargs) -> nn.Module:
+        if not self.in_memory():
+            self.load_model()
+        return super().to(*args, **kwargs)
+
+    def __call__(self, *args, **kwargs):
+        if not self.is_in_memory():
+            self.load()
+        return super().__call__(*args, **kwargs)
+
+
+class PyannoteModel(LazyModel):
+    def __init__(self, model_info, hf_token: Union[Text, bool, None] = True):
+        super().__init__()
+        self.model_info = model_info
+        self.hf_token = hf_token
+        self._model: Optional[nn.Module] = None
+
+    @property
+    def model(self) -> Optional[nn.Module]:
+        return self._model
+
+    def load(self):
+        """Load model to memory"""
+        if not self.is_in_memory():
+            self._model = pyannote_loader.get_model(self.model_info, self.hf_token)
+
+
+class PyannoteSegmentationModel(PyannoteModel):
+    def get_sample_rate(self) -> int:
+        self.load()
+        return self.model.audio.sample_rate
+
+    def get_duration(self) -> float:
+        self.load()
+        return self.model.specifications.duration
+
+    def forward(self, waveform: torch.Tensor) -> torch.Tensor:
+        return self.model(waveform)
+
+
+class PyannoteEmbeddingModel(PyannoteModel):
+    def forward(
+        self,
+        waveform: torch.Tensor,
+        weights: Optional[torch.Tensor] = None,
+    ) -> torch.Tensor:
+        return self.model(waveform, weights=weights)
 
 
 class SegmentationModel(nn.Module):
     """
     Minimal interface for a segmentation model.
     """
+    def __init__(self, model: LazyModel):
+        super().__init__()
+        self.model = model
 
     @staticmethod
     def from_pyannote(model, use_hf_token: Union[Text, bool, None] = True) -> 'SegmentationModel':
@@ -34,34 +100,17 @@ class SegmentationModel(nn.Module):
         wrapper: SegmentationModel
         """
         assert _has_pyannote, "No pyannote.audio installation found"
-
-        class PyannoteSegmentationModel(SegmentationModel):
-            def __init__(self, pyannote_model, token: Union[Text, bool, None] = True):
-                super().__init__()
-                self.model = pyannote.get_model(pyannote_model, token)
-
-            def get_sample_rate(self) -> int:
-                return self.model.audio.sample_rate
-
-            def get_duration(self) -> float:
-                return self.model.specifications.duration
-
-            def __call__(self, waveform: torch.Tensor) -> torch.Tensor:
-                return self.model(waveform)
-
-        return PyannoteSegmentationModel(model, use_hf_token)
+        return SegmentationModel(PyannoteSegmentationModel(model, use_hf_token))
 
     def get_sample_rate(self) -> int:
-        """Return the sample rate expected for model inputs"""
-        raise NotImplementedError
+        return self.model.get_sample_rate()
 
     def get_duration(self) -> float:
[…]
