### 2023-03-10 4b744ed
Refactoring of diart.models to ease custom model usage
