### 2023-03-10 cfb2a9f
Refactoring of diart.models to ease custom model usage
