+++ src/diart/models.py
@@ -1,4 +1,4 @@
-from typing import Optional, Text, Union
+from typing import Optional, Text, Union, Callable
 
 import torch
 import torch.nn as nn
@@ -10,77 +10,43 @@ except ImportError:
     _has_pyannote = False
 
 
-class LazyModel(nn.Module):
-    @property
-    def model(self) -> Optional[nn.Module]:
-        raise NotImplementedError
+class PyannoteLoader:
+    def __init__(self, model_info, hf_token: Union[Text, bool, None] = True):
+        super().__init__()
+        self.model_info = model_info
+        self.hf_token = hf_token
 
-    def load(self):
-        """Load model to memory"""
-        raise NotImplementedError
+    def __call__(self) -> nn.Module:
+        return pyannote_loader.get_model(self.model_info, self.hf_token)
+
+
+class LazyModel(nn.Module):
+    def __init__(self, loader: Callable[[], nn.Module]):
+        super().__init__()
+        self.get_model = loader
+        self.model: Optional[nn.Module] = None
 
     def is_in_memory(self) -> bool:
         """Return whether the model has been loaded into memory"""
         return self.model is not None
 
+    def load(self):
+        if not self.is_in_memory():
+            self.model = self.get_model()
+
     def to(self, *args, **kwargs) -> nn.Module:
-        if not self.in_memory():
-            self.load_model()
+        self.load()
         return super().to(*args, **kwargs)
 
     def __call__(self, *args, **kwargs):
-        if not self.is_in_memory():
-            self.load()
+        self.load()
         return super().__call__(*args, **kwargs)
 
 
-class PyannoteModel(LazyModel):
-    def __init__(self, model_info, hf_token: Union[Text, bool, None] = True):
-        super().__init__()
-        self.model_info = model_info
-        self.hf_token = hf_token
-        self._model: Optional[nn.Module] = None
-
-    @property
-    def model(self) -> Optional[nn.Module]:
-        return self._model
-
-    def load(self):
-        """Load model to memory"""
-        if not self.is_in_memory():
-            self._model = pyannote_loader.get_model(self.model_info, self.hf_token)
-
-
-class PyannoteSegmentationModel(PyannoteModel):
-    def get_sample_rate(self) -> int:
-        self.load()
-        return self.model.audio.sample_rate
-
-    def get_duration(self) -> float:
-        self.load()
-        return self.model.specifications.duration
-
-    def forward(self, waveform: torch.Tensor) -> torch.Tensor:
-        return self.model(waveform)
-
-
-class PyannoteEmbeddingModel(PyannoteModel):
-    def forward(
-        self,
-        waveform: torch.Tensor,
-        weights: Optional[torch.Tensor] = None,
-    ) -> torch.Tensor:
-        return self.model(waveform, weights=weights)
-
-
-class SegmentationModel(nn.Module):
+class SegmentationModel(LazyModel):
     """
     Minimal interface for a segmentation model.
     """
-    def __init__(self, model: LazyModel):
-        super().__init__()
-        self.model = model
-
     @staticmethod
     def from_pyannote(model, use_hf_token: Union[Text, bool, None] = True) -> 'SegmentationModel':
         """
@@ -100,13 +66,15 @@ class SegmentationModel(nn.Module):
         wrapper: SegmentationModel
         """
         assert _has_pyannote, "No pyannote.audio installation found"
-        return SegmentationModel(PyannoteSegmentationModel(model, use_hf_token))
+        return PyannoteSegmentationModel(model, use_hf_token)
 
-    def get_sample_rate(self) -> int:
-        return self.model.get_sample_rate()
+    @property
[…]
