### 2023-03-10 4a13adf
Minor refactoring
