### 2023-03-10 1cc7143
Add torchvision in requirements to avoid warning
