+++ setup.cfg
@@ -29,6 +29,7 @@ install_requires=
     tqdm>=4.64.0
     pandas>=1.4.2
     torch>=1.12.1
+    torchvision>=0.14.0
     torchaudio>=0.12.1,<1.0
     pyannote.audio>=2.1.1
     pyannote.core>=4.5
