### 2023-03-10 a9e1eea
Simplify installation of requirements in README
