+++ setup.cfg
@@ -35,7 +35,7 @@ install_requires=
     pyannote.database>=4.1.1
     pyannote.metrics>=3.2
     optuna>=2.10
-    websockets>=10.3
+    websocket-server>=0.6.4
     rich>=12.5.1
 
 [options.packages.find]
