+++ requirements.txt
@@ -13,5 +13,5 @@ pyannote.core>=4.5
 pyannote.database>=4.1.1
 pyannote.metrics>=3.2
 optuna>=2.10
-websockets>=10.3
+websocket-server>=0.6.4
 rich>=12.5.1
