### 2023-03-13 5c7b9d0
Add diart.serve script to host a pipeline on a websocket server
