+++ src/diart/sources.py
@@ -1,7 +1,7 @@
 import base64
 from pathlib import Path
 from queue import SimpleQueue
-from typing import Text, Optional, AnyStr, Dict, Any
+from typing import Text, Optional, AnyStr, Dict, Any, Union
 
 import numpy as np
 import sounddevice as sd
@@ -158,23 +158,42 @@ class WebSocketAudioSource(AudioSource):
     ----------
     sample_rate: int
         Sample rate of the chunks emitted.
-    host: Text | None
-        The host to run the websocket server. Defaults to ``None`` (all interfaces).
+    host: Text
+        The host to run the websocket server.
+        Defaults to 127.0.0.1.
     port: int
-        The port to run the websocket server. Defaults to 7007.
+        The port to run the websocket server.
+        Defaults to 7007.
+    key: Text | Path | None
+        Path to a key if using SSL.
+        Defaults to no key.
+    certificate: Text | Path | None
+        Path to a certificate if using SSL.
+        Defaults to no certificate.
     """
-    def __init__(self, sample_rate: int, host: Optional[Text] = None, port: int = 7007):
-        name = host if host is not None and host else "localhost"
-        uri = f"{name}:{port}"
+    def __init__(
+        self,
+        sample_rate: int,
+        host: Text = "127.0.0.1",
+        port: int = 7007,
+        key: Optional[Union[Text, Path]] = None,
+        certificate: Optional[Union[Text, Path]] = None
+    ):
         # FIXME sample_rate is not being used, this can be confusing and lead to incompatibilities.
         #  I would prefer the client to send a JSON with data and sample rate, then resample if needed
-        super().__init__(uri, sample_rate)
+        super().__init__(f"{host}:{port}", sample_rate)
         self.client: Optional[Dict[Text, Any]] = None
-        self.server = WebsocketServer(host, port=port)
+        self.server = WebsocketServer(host, port, key=key, cert=certificate)
         self.server.set_fn_message_received(self._on_message_received)
 
-    def _on_message_received(self, client: Dict[Text, Any], server: WebsocketServer, message: AnyStr):
-        if self.client is None:
+    def _on_message_received(
+        self,
+        client: Dict[Text, Any],
+        server: WebsocketServer,
+        message: AnyStr,
+    ):
+        # Only one client at a time is allowed
+        if self.client is None or self.client["id"] != client["id"]:
             self.client = client
         # Decode chunk encoded in base64
         byte_samples = base64.decodebytes(message.encode("utf-8"))
@@ -188,6 +207,7 @@ class WebSocketAudioSource(AudioSource):
         self.server.run_forever()
 
     def close(self):
+        """Close the websocket server"""
         if self.server is not None:
             self.stream.on_completed()
             self.server.shutdown_gracefully()
