### 2023-03-13 e4acabc
Add diart.serve script to host a pipeline on a websocket server
