+++ setup.cfg
@@ -46,3 +46,4 @@ console_scripts=
     diart.stream=diart.stream:run
     diart.benchmark=diart.benchmark:run
     diart.tune=diart.tune:run
+    diart.serve=diart.serve:run
