### 2023-03-13 08801ab
Add diart.client script
