+++ setup.cfg
@@ -36,6 +36,7 @@ install_requires=
     pyannote.metrics>=3.2
     optuna>=2.10
     websocket-server>=0.6.4
+    websocket-client>=0.58.0
     rich>=12.5.1
 
 [options.packages.find]
@@ -47,3 +48,4 @@ console_scripts=
     diart.benchmark=diart.benchmark:run
     diart.tune=diart.tune:run
     diart.serve=diart.serve:run
+    diart.client=diart.client:run
