### 2023-03-13 238f17e
Add diart.client script
