### 2023-03-13 ca71227
Create module diart.console to host all CLI scripts
