+++ setup.cfg
@@ -44,8 +44,8 @@ where=src
 
 [options.entry_points]
 console_scripts=
-    diart.stream=diart.stream:run
-    diart.benchmark=diart.benchmark:run
-    diart.tune=diart.tune:run
-    diart.serve=diart.serve:run
-    diart.client=diart.client:run
+    diart.stream=diart.console.stream:run
+    diart.benchmark=diart.console.benchmark:run
+    diart.tune=diart.console.tune:run
+    diart.serve=diart.console.serve:run
+    diart.client=diart.console.client:run
