### 2023-03-13 e4fbef4
Update websocket section in README
