### 2023-03-16 #5266 @editorialbot
[REVIEW]: Diart: A Python Library for Real-Time Speaker Diarization
**Submitting author:** @juanmc2005 (<a href="http://orcid.org/0000-0002-5035-147X">Juan Manuel Coria</a>)
**Repository:** https://github.com/juanmc2005/StreamingSpeakerDiarization
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.9.1
