### 2023-03-25 fe698eb
Add sharper logo and demo gif
