### 2023-03-25 37af28c
Reduce logo size
