### 2023-03-25 11f185a
Update logo and demo gif (#137)
