### 2023-03-25 69f3160
Shift timestamps if left padding was added to a file
