+++ setup.cfg
@@ -1,10 +1,10 @@
 [metadata]
 name=diart
-version=0.6.0
+version=0.7.0
 author=Juan Manuel Coria
 description=Speaker diarization in real time
 long_description=file: README.md
-long_description_content_type = text/markdown
+long_description_content_type=text/markdown
 keywords=speaker diarization, streaming, online, real time, rxpy
 url=https://github.com/juanmc2005/StreamingSpeakerDiarization
 license=MIT
