+++ README.md
@@ -1,7 +1,7 @@
 <br/>
 
 <p align="center">
-<img width="40%" src="/logo.jpg" title="Logo" />
+<img width="50%" src="/logo.jpg" title="Logo" />
 </p>
 
 <p align="center">
@@ -15,40 +15,40 @@
 
 <div align="center">
   <h4>
-    <a href="#installation">
-      Installation
+    <a href="#-installation">
+      💾 Installation
     </a>
     <span> | </span>
-    <a href="#stream-audio">
-      Stream audio
+    <a href="#%EF%B8%8F-stream-audio">
+      🎙️ Stream audio
     </a>
     <span> | </span>
-    <a href="#custom-models">
-      Custom models
+    <a href="#-custom-models">
+      🤖 Custom models
     </a>
     <span> | </span>
-    <a href="#tune-hyper-parameters">
-      Tune hyper-parameters
+    <a href="#-tune-hyper-parameters">
+      📈 Tune hyper-parameters
+    </a>
+    <br />
+    <a href="#--build-pipelines">
+      🧠🔗 Build pipelines
     </a>
     <span> | </span>
-    <a href="#build-pipelines">
-      Build pipelines
-    </a>
-    <br/>
-    <a href="#websockets">
-      WebSockets
+    <a href="#-websockets">
+      🌐 WebSockets
     </a>
     <span> | </span>
-    <a href="#powered-by-research">
-      Research
+    <a href="#-powered-by-research">
+      🔬 Research
     </a>
     <span> | </span>
-    <a href="#citation">
-      Citation
+    <a href="#-citation">
+      📗 Citation
     </a>
     <span> | </span>
-    <a href="#reproducibility">
-      Reproducibility
+    <a href="#-reproducibility">
+      👨‍💻 Reproducibility
     </a>
   </h4>
 </div>
@@ -59,7 +59,7 @@
 <img width="100%" src="/demo.gif" title="Real-time diarization example" />
 </p>
 
-## Installation
+## 💾 Installation
 
 1) Create environment:
 
@@ -79,7 +79,7 @@ conda install portaudio pysoundfile ffmpeg -c conda-forge
 pip install diart
 ```
 
-### Get access to pyannote models
+### Get access to 🎹 pyannote models
 
 By default, diart is based on [pyannote.audio](https://github.com/pyannote/pyannote-audio) models stored in the [huggingface](https://huggingface.co/) hub.
 To allow diart to use them, you need to follow these steps:
@@ -88,7 +88,7 @@ To allow diart to use them, you need to follow these steps:
 2) [Accept user conditions](https://huggingface.co/pyannote/embedding) for the `pyannote/embedding` model
 3) Install [huggingface-cli](https://huggingface.co/docs/huggingface_hub/quick-start#install-the-hub-library) and [log in](https://huggingface.co/docs/huggingface_hub/quick-start#login) with your user access token (or provide it manually in diart CLI or API).
 
-## Stream audio
+## 🎙️ Stream audio
 
 ### From the command line
 
@@ -127,7 +127,7 @@ prediction = inference()
 
 For inference and evaluation on a dataset we recommend to use `Benchmark` (see notes on [reproducibility](#reproducibility)).
 
-## Custom models
+## 🤖 Custom models
 
 Third-party models can be integrated seamlessly by subclassing `SegmentationModel` and `EmbeddingModel` (which are PyTorch `Module` subclasses):
 
@@ -178,7 +178,7 @@ inference = RealTimeInference(pipeline, mic)
 prediction = inference()
 ```
 
-## Tune hyper-parameters
+## 📈 Tune hyper-parameters
 
[…]
