### 2023-04-19 ade7fac
New feature: streaming voice activity detection. Pipeline name changes
