### 2023-04-19 bca2873
New feature: streaming voice activity detection. Pipeline name changes
