+++ src/diart/inference.py
@@ -4,32 +4,33 @@ from pathlib import Path
 from traceback import print_exc
 from typing import Union, Text, Optional, Callable, Tuple, List
 
-import diart.operators as dops
-import diart.sources as src
 import numpy as np
 import pandas as pd
 import rx
 import rx.operators as ops
 import torch
-from diart import utils
-from diart.blocks import BasePipeline, Resample, BasePipelineConfig
-from diart.progress import ProgressBar, RichProgressBar, TQDMProgressBar
-from diart.sinks import DiarizationPredictionAccumulator, RealTimePlot, WindowClosedException
 from pyannote.core import Annotation, SlidingWindowFeature
 from pyannote.database.util import load_rttm
-from pyannote.metrics.diarization import DiarizationErrorRate
+from pyannote.metrics.base import BaseMetric
 from rx.core import Observer
 from tqdm import tqdm
 
+from . import blocks
+from . import operators as dops
+from . import sources as src
+from . import utils
+from .progress import ProgressBar, RichProgressBar, TQDMProgressBar
+from .sinks import PredictionAccumulator, StreamingPlot, WindowClosedException
 
-class RealTimeInference:
+
+class StreamingInference:
     """Performs inference in real time given a pipeline and an audio source.
     Streams an audio source to an online speaker diarization pipeline.
     It allows users to attach a chain of operations in the form of hooks.
 
     Parameters
     ----------
-    pipeline: BasePipeline
+    pipeline: StreamingPipeline
         Configured speaker diarization pipeline.
     source: AudioSource
         Audio source to be read and streamed.
@@ -52,7 +53,7 @@ class RealTimeInference:
     """
     def __init__(
         self,
-        pipeline: BasePipeline,
+        pipeline: blocks.StreamingPipeline,
         source: src.AudioSource,
         batch_size: int = 1,
         do_profile: bool = True,
@@ -66,7 +67,7 @@ class RealTimeInference:
         self.do_profile = do_profile
         self.do_plot = do_plot
         self.show_progress = show_progress
-        self.accumulator = DiarizationPredictionAccumulator(self.source.uri)
+        self.accumulator = PredictionAccumulator(self.source.uri)
         self.unit = "chunk" if self.batch_size == 1 else "batch"
         self._observers = []
 
@@ -102,7 +103,7 @@ class RealTimeInference:
                   f"but pipeline's is {sample_rate}. Will resample."
             logging.warning(msg)
             self.stream = self.stream.pipe(
-                ops.map(Resample(self.source.sample_rate, sample_rate))
+                ops.map(blocks.Resample(self.source.sample_rate, sample_rate))
             )
 
         # Add rx operators to manage the inputs and outputs of the pipeline
@@ -202,7 +203,7 @@ class RealTimeInference:
                     latency=config.latency,
                     sample_rate=config.sample_rate,
                 ),
-                ops.do(RealTimePlot(config.duration, config.latency)),
+                ops.do(StreamingPlot(config.duration, config.latency)),
             )
         observable.subscribe(
             on_error=self._handle_error,
@@ -288,7 +289,7 @@ class Benchmark:
 
     def run_single(
         self,
-        pipeline: BasePipeline,
+        pipeline: blocks.StreamingPipeline,
         filepath: Path,
         progress_bar: ProgressBar,
     ) -> Annotation:
@@ -298,7 +299,7 @@ class Benchmark:
 
         Parameters
         ----------
-        pipeline: BasePipeline
+        pipeline: StreamingPipeline
             Speaker diarization pipeline to run.
         filepath: Path
             Path to the target file.
@@ -318,7 +319,7 @@ class Benchmark:
             pipeline.config.optimal_block_size(),
         )
         pipeline.set_timestamp_shift(-padding[0])
-        inference = RealTimeInference(
[…]
