### 2023-04-19 d19b044
New feature: streaming voice activity detection. Pipeline name changes
