### 2023-04-19 d80ac32
Add minor README modifications
