### 2023-04-21 8cc9925
Initial ASR implementation. Broken stuff
