+++ src/diart/blocks/diarization.py
@@ -192,7 +192,8 @@ class SpeakerDiarization(base.StreamingPipeline):
 
     def __call__(
         self,
-        waveforms: Sequence[SlidingWindowFeature]
+        waveforms: Sequence[SlidingWindowFeature],
+        **kwargs,
     ) -> Sequence[Tuple[Annotation, SlidingWindowFeature]]:
         batch_size = len(waveforms)
         msg = "Pipeline expected at least 1 input"
