+++ src/diart/blocks/base.py
@@ -87,6 +87,7 @@ class StreamingPipeline:
 
     def __call__(
         self,
-        waveforms: Sequence[SlidingWindowFeature]
+        waveforms: Sequence[SlidingWindowFeature],
+        **kwargs,
     ) -> Sequence[Tuple[Any, SlidingWindowFeature]]:
         raise NotImplementedError

[…]
