### 2023-04-21 1ae4934
First working transcription pipeline. Using diarization is possible but a bit quirky
