+++ src/diart/inference.py
@@ -2,7 +2,7 @@ import logging
 from multiprocessing import Pool, freeze_support, RLock, current_process
 from pathlib import Path
 from traceback import print_exc
-from typing import Union, Text, Optional, Callable, Tuple, List, Any
+from typing import Union, Text, Optional, Callable, Tuple, List, Any, Dict
 
 import numpy as np
 import pandas as pd
@@ -10,8 +10,6 @@ import rx
 import rx.operators as ops
 import torch
 from pyannote.core import Annotation, SlidingWindowFeature
-from pyannote.database.util import load_rttm
-from pyannote.metrics.base import BaseMetric
 from rx.core import Observer
 from tqdm import tqdm
 
@@ -19,8 +17,9 @@ from . import blocks
 from . import operators as dops
 from . import sources as src
 from . import utils
+from .metrics import Metric
 from .progress import ProgressBar, RichProgressBar, TQDMProgressBar
-from .sinks import DiarizationAccumulator, StreamingPlot, WindowClosedException
+from .sinks import StreamingPlot, WindowClosedException
 
 
 class StreamingInference:
@@ -292,7 +291,7 @@ class Benchmark:
         pipeline: blocks.StreamingPipeline,
         filepath: Path,
         progress_bar: ProgressBar,
-    ) -> Annotation:
+    ) -> Tuple[Text, Any]:
         """Run a given pipeline on a given file.
         Note that this method does NOT reset the
         state of the pipeline before execution.
@@ -329,36 +328,29 @@ class Benchmark:
             progress_bar=progress_bar,
         )
 
-        # Accumulate predictions in memory
-        pred_accumulator = DiarizationAccumulator(source.uri)
-        inference.attach_observers(pred_accumulator)
+        # Run the pipeline and concatenate predictions
+        pred = pipeline.join_predictions(inference())
 
-        # Run the pipeline on this file
-        inference()
-
-        # Extract prediction
-        pred = pred_accumulator.get_prediction()
-
+        # Write prediction to disk if required
         if self.output_path is not None:
-            with open(self.output_path / f"{source.uri}.rttm", "w") as out_file:
-                pred.write_rttm(out_file)
+            pipeline.write_prediction(source.uri, pred, self.output_path)
 
-        return pred
+        return source.uri, pred
 
     def evaluate(
         self,
-        predictions: List[Annotation],
-        metric: BaseMetric,
-    ) -> Union[pd.DataFrame, List[Annotation]]:
+        predictions: Dict[Text, Any],
+        metric: Metric,
+    ) -> Union[pd.DataFrame, Dict[Text, Any]]:
         """If a reference path was provided,
         compute the diarization error rate of a list of predictions.
 
         Parameters
         ----------
-        predictions: List[Annotation]
+        predictions: List[Any]
             Predictions to evaluate.
-        metric: BaseMetric
-            Evaluation metric from pyannote.metrics.
+        metric: Metric
+            Evaluation metric.
 
         Returns
         -------
@@ -367,23 +359,37 @@ class Benchmark:
             reference path was given. Otherwise return the same predictions.
         """
         if self.reference_path is not None:
+            # Initialize progress bar
             progress_bar = TQDMProgressBar(f"Computing {metric.name}", leave=False)
             progress_bar.create(total=len(predictions), unit="file")
             progress_bar.start()
-            for hyp in predictions:
-                ref = load_rttm(self.reference_path / f"{hyp.uri}.rttm").popitem()[1]
-                metric(ref, hyp)
+
+            # Evaluate each prediction
+            uris = []
+            for uri, pred in predictions.items():
+                ref_file = list(self.reference_path.glob(f"{uri}.*"))
+                if ref_file:
[…]
