### 2023-04-21 d8d7342
Reduce Whisper VRAM footprint (around 400Mb). Add fp16 option
