### 2023-04-21 2cfc35d
Change whisper input type based on fp16 parameter
