+++ src/diart/models.py
@@ -1,7 +1,9 @@
+import time
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Optional, Text, Union, Callable, List, Tuple, Dict
+from typing import Optional, Text, Union, Callable, List, Any
 
+import numpy as np
 import torch
 import torch.nn as nn
 from pyannote.core import Segment
@@ -14,9 +16,16 @@ except ImportError:
 
 try:
     import whisper
+    from whisper.tokenizer import get_tokenizer
     _has_whisper = True
+    DecodingResult = whisper.DecodingResult
+    DecodingOptions = whisper.DecodingOptions
+    Tokenizer = whisper.tokenizer.Tokenizer
 except ImportError:
     _has_whisper = False
+    DecodingResult = Any
+    DecodingOptions = Any
+    Tokenizer = Any
 
 
 class PyannoteLoader:
@@ -207,15 +216,12 @@ class SpeechRecognitionModel(LazyModel):
         name: Text,
         download_path: Optional[Union[Text, Path]] = None,
         in_memory: bool = False,
-        remember_transcriptions: bool = True,
         fp16: bool = False,
     ) -> 'SpeechRecognitionModel':
         msg = "No whisper-transcribed installation found. " \
               "Visit https://github.com/linto-ai/whisper-timestamped#installation to install"
         assert _has_whisper, msg
-        return WhisperSpeechRecognitionModel(
-            name, download_path, in_memory, remember_transcriptions, fp16
-        )
+        return WhisperSpeechRecognitionModel(name, download_path, in_memory, fp16)
 
     @property
     def duration(self) -> float:
@@ -245,20 +251,141 @@ class SpeechRecognitionModel(LazyModel):
         raise NotImplementedError
 
 
+class WhisperDecoder:
+    def __init__(
+        self,
+        compression_ratio_threshold: Optional[float] = 2.4,
+        logprob_threshold: Optional[float] = -1,
+    ):
+        self.compression_ratio_threshold = compression_ratio_threshold
+        self.logprob_threshold = logprob_threshold
+        self.temperatures = (0, 0.2, 0.4, 0.6, 0.8, 1)
+
+    @staticmethod
+    def get_temperature_options(initial: DecodingOptions, t: float) -> DecodingOptions:
+        t_options = {**vars(initial)}
+        if t > 0:
+            t_options.pop("beam_size", None)
+            t_options.pop("patience", None)
+        else:
+            t_options.pop("best_of", None)
+        t_options["temperature"] = t
+        return DecodingOptions(**t_options)
+
+    @staticmethod
+    def decode(
+        model,
+        batch: torch.Tensor,
+        options: DecodingOptions
+    ) -> DecodingResult:
+        return model.decode(batch, options)
+
+    def check_compression(self) -> bool:
+        return self.compression_ratio_threshold is not None
+
+    def check_logprob(self) -> bool:
+        return self.logprob_threshold is not None
+
+    def needs_fallback(self, output: DecodingResult) -> bool:
+        if self.check_compression and output.compression_ratio > self.compression_ratio_threshold:
+            # Transcription is too repetitive
+            return True
+        if self.check_logprob and output.avg_logprob < self.logprob_threshold:
+            # Average log probability is too low
+            return True
+        return False
+
+    def decode_with_fallback(
+        self,
+        model,
+        batch: torch.Tensor,
+        options: DecodingOptions,
+    ) -> DecodingResult:
+        batch_size = batch.shape[0]
+        results = [None] * batch_size
+        retry_idx = torch.ones(batch_size).type(torch.bool)
+
+        for t in self.temperatures:
+            # Transcribe with the given temperature
+            t_options = self.get_temperature_options(options, t)
[…]
