### 2023-04-22 e8196a7
Minor changes in transcription arguments
