+++ src/diart/blocks/diarization.py
@@ -206,7 +206,6 @@ class SpeakerDiarization(base.StreamingPipeline):
     def __call__(
         self,
         waveforms: Sequence[SlidingWindowFeature],
-        **kwargs,
     ) -> Sequence[Tuple[Annotation, SlidingWindowFeature]]:
         batch_size = len(waveforms)
         msg = "Pipeline expected at least 1 input"
