### 2023-04-23 07dd9ae
Greatly improve transcription pipeline by adding optional VAD
