+++ src/diart/models.py
@@ -204,7 +204,7 @@ class PyannoteEmbeddingModel(EmbeddingModel):
 
 
 @dataclass(frozen=True)
-class Transcription:
+class TranscriptionResult:
     text: Text
     chunks: List[Text]
     timestamps: List[Segment]
@@ -217,11 +217,24 @@ class SpeechRecognitionModel(LazyModel):
         download_path: Optional[Union[Text, Path]] = None,
         in_memory: bool = False,
         fp16: bool = False,
+        no_speech_threshold: float = 0.6,
+        compression_ratio_threshold: Optional[float] = 2.4,
+        logprob_threshold: Optional[float] = -1,
+        decode_with_fallback: bool = False,
     ) -> 'SpeechRecognitionModel':
         msg = "No whisper-transcribed installation found. " \
               "Visit https://github.com/linto-ai/whisper-timestamped#installation to install"
         assert _has_whisper, msg
-        return WhisperSpeechRecognitionModel(name, download_path, in_memory, fp16)
+        return WhisperSpeechRecognitionModel(
+            name,
+            download_path,
+            in_memory,
+            fp16,
+            no_speech_threshold,
+            compression_ratio_threshold,
+            logprob_threshold,
+            decode_with_fallback,
+        )
 
     @property
     def duration(self) -> float:
@@ -237,7 +250,7 @@ class SpeechRecognitionModel(LazyModel):
     def set_beam_size(self, size: Optional[int] = None):
         raise NotImplementedError
 
-    def forward(self, waveform: torch.Tensor) -> List[Transcription]:
+    def forward(self, waveform: torch.Tensor) -> List[TranscriptionResult]:
         """
         Forward pass of the speech recognition model.
 
@@ -248,7 +261,7 @@ class SpeechRecognitionModel(LazyModel):
 
         Returns
         -------
-        transcriptions: List[Transcription]
+        transcriptions: List[TranscriptionResult]
             A list of timestamped transcriptions
         """
         raise NotImplementedError
@@ -257,9 +270,11 @@ class SpeechRecognitionModel(LazyModel):
 class WhisperDecoder:
     def __init__(
         self,
+        no_speech_threshold: float = 0.6,
         compression_ratio_threshold: Optional[float] = 2.4,
         logprob_threshold: Optional[float] = -1,
     ):
+        self.no_speech_threshold = no_speech_threshold
         self.compression_ratio_threshold = compression_ratio_threshold
         self.logprob_threshold = logprob_threshold
         self.temperatures = (0, 0.2, 0.4, 0.6, 0.8, 1)
@@ -303,7 +318,24 @@ class WhisperDecoder:
         model,
         batch: torch.Tensor,
         options: DecodingOptions,
-    ) -> DecodingResult:
+    ) -> List[DecodingResult]:
+        """Transcribe batch and retry with ever-increasing
+        temperatures if the estimated quality of the transcription is not good.
+
+        Parameters
+        ----------
+        model: whisper.Whisper
+            Whisper ASR model (contains 'decode' method).
+        batch: torch.Tensor, shape (batch, channel, samples)
+            Log mel spectrogram batch.
+        options: whisper.DecodingOptions
+            Configuration to decode transcription.
+
+        Returns
+        -------
+        result: List[whisper.DecodingResult]
+            Transcription results for this batch.
+        """
         batch_size = batch.shape[0]
         results = [None] * batch_size
         retry_idx = torch.ones(batch_size).type(torch.bool)
@@ -314,26 +346,51 @@ class WhisperDecoder:
             outputs = model.decode(batch[retry_idx], t_options)
 
             # Determine which outputs need to be transcribed again
-            #  based on quality estimates
             output_idx = torch.where(retry_idx)[0]
             for idx, out in zip(output_idx, outputs):
                 results[idx] = out
[…]
