### 2023-04-23 0bf2522
Move pipelines to diart.pipelines. Add torchmetrics as a dependency
