+++ src/diart/pipelines/voice.py
@@ -1,36 +1,27 @@
 from pathlib import Path
-from typing import Optional, Tuple, Sequence, Union, Any, Text, List
+from typing import Any, Optional, Union, Sequence, Tuple, Text, List
 
 import numpy as np
 import torch
-from pyannote.core import Annotation, SlidingWindowFeature, SlidingWindow, Segment
+from pyannote.core import Annotation, Timeline, SlidingWindowFeature, SlidingWindow, Segment
 from typing_extensions import Literal
 
 from . import base
-from .aggregation import DelayedAggregation
-from .clustering import OnlineSpeakerClustering
-from .embedding import OverlapAwareSpeakerEmbedding
-from .segmentation import SpeakerSegmentation
-from .utils import Binarize
+from .hparams import HyperParameter, TauActive
+from .. import blocks
 from .. import models as m
 from .. import utils
-from ..metrics import Metric, DiarizationErrorRate
+from ..metrics import Metric, DetectionErrorRate
 
 
-class SpeakerDiarizationConfig(base.StreamingConfig):
+class VoiceActivityDetectionConfig(base.StreamingConfig):
     def __init__(
         self,
         segmentation: Optional[m.SegmentationModel] = None,
-        embedding: Optional[m.EmbeddingModel] = None,
         duration: Optional[float] = None,
         step: float = 0.5,
         latency: Optional[Union[float, Literal["max", "min"]]] = None,
         tau_active: float = 0.6,
-        rho_update: float = 0.3,
-        delta_new: float = 1,
-        gamma: float = 3,
-        beta: float = 10,
-        max_speakers: int = 20,
         merge_collar: float = 0.05,
         device: Optional[torch.device] = None,
         **kwargs,
@@ -41,15 +32,10 @@ class SpeakerDiarizationConfig(base.StreamingConfig):
             self.segmentation = m.SegmentationModel.from_pyannote("pyannote/segmentation")
 
         self._duration = duration
+        self._step = step
         self._sample_rate: Optional[int] = None
 
-        # Default embedding model is pyannote/embedding
-        self.embedding = embedding
-        if self.embedding is None:
-            self.embedding = m.EmbeddingModel.from_pyannote("pyannote/embedding")
-
         # Latency defaults to the step duration
-        self._step = step
         self._latency = latency
         if self._latency is None or self._latency == "min":
             self._latency = self._step
@@ -57,57 +43,10 @@ class SpeakerDiarizationConfig(base.StreamingConfig):
             self._latency = self._duration
 
         self.tau_active = tau_active
-        self.rho_update = rho_update
-        self.delta_new = delta_new
-        self.gamma = gamma
-        self.beta = beta
-        self.max_speakers = max_speakers
         self.merge_collar = merge_collar
-
         self.device = device
         if self.device is None:
             self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-
-    @staticmethod
-    def from_dict(data: Any) -> 'SpeakerDiarizationConfig':
-        # Check for explicit device, otherwise check for 'cpu' bool, otherwise pass None
-        device = utils.get(data, "device", None)
-        if device is None:
-            device = torch.device("cpu") if utils.get(data, "cpu", False) else None
-
-        # Instantiate models
-        hf_token = utils.parse_hf_token_arg(utils.get(data, "hf_token", True))
-        segmentation = utils.get(data, "segmentation", "pyannote/segmentation")
-        segmentation = m.SegmentationModel.from_pyannote(segmentation, hf_token)
-        embedding = utils.get(data, "embedding", "pyannote/embedding")
-        embedding = m.EmbeddingModel.from_pyannote(embedding, hf_token)
-
-        # Hyper-parameters and their aliases
-        tau = utils.get(data, "tau_active", None)
[…]
