+++ src/diart/pipelines/transcription.py
@@ -1,0 +1,184 @@
+from pathlib import Path
+from typing import Sequence, Optional, Any, Union, List, Text, Tuple
+
+import numpy as np
+import torch
+from pyannote.core import SlidingWindowFeature
+
+from . import base
+from .hparams import HyperParameter, TauActive
+from .. import blocks
+from .. import models as m
+from .. import utils
+from ..metrics import Metric, WordErrorRate
+
+
+class TranscriptionConfig(base.StreamingConfig):
+    def __init__(
+        self,
+        asr: Optional[m.SpeechRecognitionModel] = None,
+        vad: Optional[m.SegmentationModel] = None,
+        tau_active: float = 0.5,
+        duration: Optional[float] = None,
+        language: Optional[Text] = None,
+        beam_size: int = None,
+        device: Optional[torch.device] = None,
+    ):
+        self.device = device
+        if self.device is None:
+            self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+
+        # Default ASR model is Whisper small (244M parameters)
+        self.asr = asr
+        if self.asr is None:
+            self.asr = m.SpeechRecognitionModel.from_whisper("small")
+        self.asr.set_language(language)
+        self.asr.set_beam_size(beam_size)
+
+        self.vad = vad
+        self.tau_active = tau_active
+
+        self._duration = duration
+        self._sample_rate: Optional[int] = None
+
+    @property
+    def duration(self) -> float:
+        if self._duration is None:
+            self._duration = self.asr.duration
+        return self._duration
+
+    @property
+    def step(self) -> float:
+        return self.duration
+
+    @property
+    def latency(self) -> float:
+        return self.duration
+
+    @property
+    def sample_rate(self) -> int:
+        if self._sample_rate is None:
+            self._sample_rate = self.asr.sample_rate
+        return self._sample_rate
+
+    @staticmethod
+    def from_dict(data: Any) -> 'TranscriptionConfig':
+        # Check for explicit device, otherwise check for 'cpu' bool, otherwise pass None
+        device = utils.get(data, "device", None)
+        if device is None:
+            device = torch.device("cpu") if utils.get(data, "cpu", False) else None
+        return TranscriptionConfig(
+            asr=utils.get(data, "asr", None),
+            vad=utils.get(data, "vad", None),
+            tau_active=utils.get(data, "tau_active", None),
+            duration=utils.get(data, "duration", None),
+            language=utils.get(data, "language", None),
+            beam_size=utils.get(data, "beam_size", None),
+            device=device,
+        )
+
+
+class Transcription(base.StreamingPipeline):
+    def __init__(self, config: Optional[TranscriptionConfig] = None):
+        self._config = TranscriptionConfig() if config is None else config
+        self.asr = blocks.SpeechRecognition(self.config.asr, self.config.device)
+        self.segmentation = None
+        if self.config.vad is not None:
+            self.segmentation = blocks.SpeakerSegmentation(self.config.vad, self.config.device)
+
+    @staticmethod
+    def get_config_class() -> type:
+        return TranscriptionConfig
+
+    @staticmethod
+    def suggest_metric() -> Metric:
+        return WordErrorRate()
+
+    @staticmethod
+    def hyper_parameters() -> Sequence[HyperParameter]:
+        return [TauActive]
+
+    @property
+    def config(self) -> TranscriptionConfig:
+        return self._config
+
+    def reset(self):
+        # No internal state. Nothing to do
+        pass
+
+    def set_timestamp_shift(self, shift: float):
+        # No timestamped output. Nothing to do
+        pass
+
+    def join_predictions(self, predictions: List[Text]) -> Text:
+        return "\n".join(predictions)
+
[…]
