### 2023-04-23 42fe5f7
Add websocket compatibility to transcription pipeline
