+++ src/diart/utils.py
@@ -92,6 +92,12 @@ def get_padding_right(latency: float, step: float) -> float:
     return latency - step
 
 
+def serialize_prediction(value: Union[Annotation, Text]) -> Text:
+    if isinstance(value, Annotation):
+        return value.to_rttm()
+    return value
+
+
 def visualize_feature(duration: Optional[float] = None):
     def apply(feature: SlidingWindowFeature):
         if duration is None:
