+++ src/diart/pipelines/transcription.py
@@ -4,11 +4,13 @@ from typing import Sequence, Optional, Any, Union, List, Text, Tuple
 import numpy as np
 import torch
 from pyannote.core import SlidingWindowFeature
+from rx.core import Observer
 
 from . import base
 from .hparams import HyperParameter, TauActive
 from .. import blocks
 from .. import models as m
+from .. import sinks
 from .. import utils
 from ..metrics import Metric, WordErrorRate
 
@@ -17,9 +19,9 @@ class TranscriptionConfig(base.StreamingConfig):
     def __init__(
         self,
         asr: Optional[m.SpeechRecognitionModel] = None,
-        vad: Optional[m.SegmentationModel] = None,
+        segmentation: Optional[m.SegmentationModel] = None,
         tau_active: float = 0.5,
-        duration: Optional[float] = None,
+        duration: Optional[float] = 3,
         language: Optional[Text] = None,
         beam_size: int = None,
         device: Optional[torch.device] = None,
@@ -35,7 +37,7 @@ class TranscriptionConfig(base.StreamingConfig):
         self.asr.set_language(language)
         self.asr.set_beam_size(beam_size)
 
-        self.vad = vad
+        self.segmentation = segmentation
         self.tau_active = tau_active
 
         self._duration = duration
@@ -67,11 +69,27 @@ class TranscriptionConfig(base.StreamingConfig):
         device = utils.get(data, "device", None)
         if device is None:
             device = torch.device("cpu") if utils.get(data, "cpu", False) else None
+
+        # Default ASR model is Whisper small (244M parameters)
+        whisper_size = utils.get(data, "whisper", "small")
+        asr = m.SpeechRecognitionModel.from_whisper(whisper_size)
+
+        # No VAD segmentation by default
+        segmentation = utils.get(data, "segmentation", None)
+        if segmentation is not None:
+            hf_token = utils.parse_hf_token_arg(utils.get(data, "hf_token", True))
+            segmentation = m.SegmentationModel.from_pyannote(segmentation, hf_token)
+
+        # Tau hyper-parameter and its alias
+        tau = utils.get(data, "tau_active", None)
+        if tau is None:
+            tau = utils.get(data, "tau", 0.5)
+
         return TranscriptionConfig(
-            asr=utils.get(data, "asr", None),
-            vad=utils.get(data, "vad", None),
-            tau_active=utils.get(data, "tau_active", None),
-            duration=utils.get(data, "duration", None),
+            asr=asr,
+            segmentation=segmentation,
+            tau_active=tau,
+            duration=utils.get(data, "duration", 3),
             language=utils.get(data, "language", None),
             beam_size=utils.get(data, "beam_size", None),
             device=device,
@@ -83,8 +101,8 @@ class Transcription(base.StreamingPipeline):
         self._config = TranscriptionConfig() if config is None else config
         self.asr = blocks.SpeechRecognition(self.config.asr, self.config.device)
         self.segmentation = None
-        if self.config.vad is not None:
-            self.segmentation = blocks.SpeakerSegmentation(self.config.vad, self.config.device)
+        if self.config.segmentation is not None:
+            self.segmentation = blocks.SpeakerSegmentation(self.config.segmentation, self.config.device)
 
     @staticmethod
     def get_config_class() -> type:
@@ -93,6 +111,10 @@ class Transcription(base.StreamingPipeline):
     @staticmethod
     def suggest_metric() -> Metric:
         return WordErrorRate()
+
+    @staticmethod
+    def suggest_writer(uri: Text, output_dir: Union[Text, Path]) -> Observer:
+        return sinks.TextWriter(Path(output_dir) / f"{uri}.txt")
 
     @staticmethod
     def hyper_parameters() -> Sequence[HyperParameter]:
