### 2023-04-23 49616e5
Transcription pipeline is now fully compatible with diart.stream
