+++ src/diart/sinks.py
@@ -1,12 +1,14 @@
+import re
 from pathlib import Path
-from typing import Union, Text, Optional, Tuple, Any
+from typing import Union, Text, Optional, Tuple, Any, List
 
 import matplotlib.pyplot as plt
-from pyannote.core import Annotation, Segment, SlidingWindowFeature, notebook
+import numpy as np
+import rich
+from pyannote.core import Annotation, Segment, SlidingWindowFeature, SlidingWindow, notebook
 from pyannote.database.util import load_rttm
 from pyannote.metrics.diarization import DiarizationErrorRate
 from rx.core import Observer
-from typing_extensions import Literal
 
 
 class WindowClosedException(Exception):
@@ -99,25 +101,59 @@ class DiarizationAccumulator(Observer):
         self.patch()
 
 
+class RichScreen(Observer):
+    def __init__(self, speaker_colors: Optional[List[Text]] = None):
+        super().__init__()
+        self.colors = speaker_colors
+        if self.colors is None:
+            self.colors = [
+                "bright_red", "bright_blue", "bright_green", "orange3", "deep_pink1",
+                "yellow2", "magenta", "cyan", "bright_magenta", "dodger_blue2"
+            ]
+        self.num_colors = len(self.colors)
+
+    def on_next(self, value: Union[Tuple, Text]):
+        prediction = _extract_prediction(value)
+        # Extract speakers
+        speakers = sorted(re.findall(r'\[.*?]', prediction))
+        # Colorize based on speakers
+        colorized = prediction
+        for i, speaker in enumerate(speakers):
+            colorized = colorized.replace(speaker, f"[{self.colors[i % self.num_colors]}]")
+        # Print result
+        rich.print(colorized)
+
+
 class StreamingPlot(Observer):
     def __init__(
         self,
         duration: float,
+        step: float,
         latency: float,
-        visualization: Literal["slide", "accumulate"] = "slide",
+        sample_rate: float,
         reference: Optional[Union[Path, Text]] = None,
+        patch_collar: float = 0.05,
     ):
         super().__init__()
-        assert visualization in ["slide", "accumulate"]
-        self.visualization = visualization
         self.reference = reference
         if self.reference is not None:
             self.reference = list(load_rttm(reference).values())[0]
         self.window_duration = duration
+        self.window_step = step
         self.latency = latency
+        self.sample_rate = sample_rate
+        self.patch_collar = patch_collar
+
+        self.num_window_samples = int(np.rint(self.window_duration * self.sample_rate))
+        self.num_step_samples = int(np.rint(self.window_step * self.sample_rate))
+        self.audio_resolution = 1 / self.sample_rate
+
         self.figure, self.axs, self.num_axs = None, None, -1
         # This flag allows to catch the matplotlib window closed event and make the next call stop iterating
         self.window_closed = False
+        self.real_time = 0
+        self.pred_buffer, self.audio_buffer = None, None
+        self.next_sample = 0
 
     def _on_window_closed(self, event):
         self.window_closed = True
@@ -139,21 +175,63 @@ class StreamingPlot(Observer):
         for i in range(self.num_axs):
             self.axs[i].clear()
 
-    def get_plot_bounds(self, real_time: float) -> Segment:
-        start_time = 0
-        end_time = real_time - self.latency
-        if self.visualization == "slide":
-            start_time = max(0., end_time - self.window_duration)
+    def get_plot_bounds(self) -> Segment:
+        end_time = self.real_time - self.latency
+        start_time = max(0., end_time - self.window_duration)
         return Segment(start_time, end_time)
 
     def on_next(
         self,
-        values: Tuple[Annotation, SlidingWindowFeature, float]
+        values: Tuple[Annotation, SlidingWindowFeature]
     ):
         if self.window_closed:
[…]
