+++ src/diart/console/serve.py
@@ -51,7 +51,6 @@ def run():
         audio_source,
         batch_size=1,
         do_profile=False,
-        do_plot=False,
         show_progress=True,
     )
 

[…]
