+++ src/diart/inference.py
@@ -18,22 +18,21 @@ from . import operators as dops
 from . import sources as src
 from . import utils
 from .metrics import Metric
-from .pipelines import StreamingPipeline, StreamingConfig
+from .pipelines import Pipeline, PipelineConfig
 from .progress import ProgressBar, RichProgressBar, TQDMProgressBar
 from .sinks import WindowClosedException
 
 
 class StreamingInference:
-    """Performs inference in real time given a pipeline and an audio source.
-    Streams an audio source to an online speaker diarization pipeline.
-    It allows users to attach a chain of operations in the form of hooks.
+    """Performs streaming inference given a pipeline and an audio source.
+    Side-effect hooks and observers can also be attached for customized behavior.
 
     Parameters
     ----------
-    pipeline: StreamingPipeline
-        Configured speaker diarization pipeline.
+    pipeline: Pipeline
+        A pipeline.
     source: AudioSource
-        Audio source to be read and streamed.
+        Audio source to read and stream.
     batch_size: int
         Number of inputs to send to the pipeline at once.
         Defaults to 1.
@@ -50,7 +49,7 @@ class StreamingInference:
     """
     def __init__(
         self,
-        pipeline: StreamingPipeline,
+        pipeline: Pipeline,
         source: src.AudioSource,
         batch_size: int = 1,
         do_profile: bool = True,
@@ -62,7 +61,6 @@ class StreamingInference:
         self.batch_size = batch_size
         self.do_profile = do_profile
         self.show_progress = show_progress
-        self.unit = "chunk" if self.batch_size == 1 else "batch"
         self._observers = []
         self._predictions = []
 
@@ -84,11 +82,11 @@ class StreamingInference:
             self._pbar.create(
                 total=self.num_chunks,
                 description=f"Streaming {self.source.uri}",
-                unit=self.unit
+                unit="chunk"
             )
 
         # Initialize chronometer for profiling
-        self._chrono = utils.Chronometer(self.unit, self._pbar)
+        self._chrono = utils.Chronometer("batch", self._pbar)
 
         self.stream = self.source.stream
 
@@ -198,21 +196,21 @@ class StreamingInference:
 
 class Benchmark:
     """
-    Run an online speaker diarization pipeline on a set of audio files in batches.
+    Run a pipeline on a set of audio files in batches.
     Write predictions to a given output directory.
 
-    If the reference is given, calculate the average diarization error rate.
+    If the reference is given, compute the average performance metric.
 
     Parameters
     ----------
     speech_path: Text or Path
         Directory with audio files.
     reference_path: Text, Path or None
-        Directory with reference RTTM files (same names as audio files).
-        If None, performance will not be calculated.
+        Directory with reference files (same names as audio files with different extension).
+        If None, performance will not be computed.
         Defaults to None.
-    output_path: Text, Path or None
-        Output directory to store predictions in RTTM format.
+    output_path: Optional[Text | Path]
+        Output directory to store predictions.
         If None, predictions will not be written to disk.
         Defaults to None.
     show_progress: bool
@@ -222,12 +220,7 @@ class Benchmark:
         Whether to print a performance report to stdout.
         Defaults to True.
     batch_size: int
-        Inference batch size.
-        If < 2, then it will run in real time.
-        If >= 2, then it will pre-calculate segmentation and
-        embeddings, running the rest in real time.
-        The performance between this two modes does not differ.
-        Defaults to 32.
+        Inference batch size. Defaults to 32.
     """
     def __init__(
         self,
@@ -252,7 +245,7 @@ class Benchmark:
 
         self.output_path = output_path
         if self.output_path is not None:
-            self.output_path = Path(output_path).expanduser()
[…]
