+++ src/diart/console/tune.py
@@ -16,10 +16,16 @@ def run():
                         help="Directory with RTTM files CONVERSATION.rttm. Names must match audio files")
     parser.add_argument("--pipeline", default="SpeakerDiarization", type=str,
                         help="Class of the pipeline to optimize. Defaults to 'SpeakerDiarization'")
+    parser.add_argument("--whisper", default="small", type=str,
+                        help=f"Whisper model for transcription pipeline. Defaults to 'small'")
+    parser.add_argument("--language", default="en", type=str,
+                        help=f"Transcribe in this language. Defaults to 'en' (English)")
     parser.add_argument("--segmentation", default="pyannote/segmentation", type=str,
                         help=f"{argdoc.SEGMENTATION}. Defaults to pyannote/segmentation")
     parser.add_argument("--embedding", default="pyannote/embedding", type=str,
                         help=f"{argdoc.EMBEDDING}. Defaults to pyannote/embedding")
+    parser.add_argument("--duration", default=5, type=float,
+                        help=f"Duration of the sliding window (in seconds). Default value depends on the pipeline")
     parser.add_argument("--step", default=0.5, type=float, help=f"{argdoc.STEP}. Defaults to 0.5")
     parser.add_argument("--latency", default=0.5, type=float, help=f"{argdoc.LATENCY}. Defaults to 0.5")
     parser.add_argument("--tau", default=0.5, type=float, help=f"{argdoc.TAU}. Defaults to 0.5")
@@ -32,10 +38,12 @@ def run():
     parser.add_argument("--cpu", dest="cpu", action="store_true",
                         help=f"{argdoc.CPU}. Defaults to GPU if available, CPU otherwise")
     parser.add_argument("--hparams", nargs="+", default=("tau_active", "rho_update", "delta_new"),
-                        help="Hyper-parameters to optimize. Must match names in `PipelineConfig`. Defaults to tau_active, rho_update and delta_new")
+                        help="Hyper-parameters to optimize. Must match names in `PipelineConfig`. "
+                             "Defaults to tau_active, rho_update and delta_new")
     parser.add_argument("--num-iter", default=100, type=int, help="Number of optimization trials")
     parser.add_argument("--storage", type=str,
-                        help="Optuna storage string. If provided, continue a previous study instead of creating one. The database name must match the study name")
+                        help="Optuna storage string. If provided, continue a previous study instead of creating one. "
+                             "The database name must match the study name")
     parser.add_argument("--output", type=str, help="Working directory")
     parser.add_argument("--hf-token", default="true", type=str,
                         help=f"{argdoc.HF_TOKEN}. Defaults to 'true' (required by pyannote)")
