+++ src/diart/pipelines/__init__.py
@@ -1,4 +1,4 @@
-from .base import StreamingPipeline, StreamingConfig
+from .base import Pipeline, PipelineConfig
 from .diarization import SpeakerDiarization, SpeakerDiarizationConfig
 from .transcription import Transcription, TranscriptionConfig
 from .voice import VoiceActivityDetection, VoiceActivityDetectionConfig

[…]
