### 2023-04-24 cc97b75
Rename base pipeline and config objects
