### 2023-04-24 6609e3c
Rename base pipeline and config objects
