### 2023-04-24 569c68f
Rename base pipeline and config objects
