### 2023-04-24 a16bb5c
Add initial implementation of SpeakerAwareTranscription
