+++ src/diart/pipelines/__init__.py
@@ -1,4 +1,5 @@
 from .base import Pipeline, PipelineConfig
 from .diarization import SpeakerDiarization, SpeakerDiarizationConfig
+from .speaker_transcription import SpeakerAwareTranscription, SpeakerAwareTranscriptionConfig
 from .transcription import Transcription, TranscriptionConfig
 from .voice import VoiceActivityDetection, VoiceActivityDetectionConfig
