+++ src/diart/pipelines/speaker_transcription.py
@@ -1,0 +1,316 @@
+from pathlib import Path
+from typing import Any, Optional, Union, Sequence, Tuple, Text, List
+
+import numpy as np
+import torch
+from diart.metrics import Metric
+from pyannote.core import SlidingWindowFeature, SlidingWindow, Annotation, Segment
+from rx.core import Observer
+from typing_extensions import Literal
+
+from .base import Pipeline, PipelineConfig
+from .diarization import SpeakerDiarization, SpeakerDiarizationConfig
+from .hparams import HyperParameter, TauActive, RhoUpdate, DeltaNew
+from .. import models as m
+from .. import sinks
+from .. import blocks
+from .. import utils
+from ..metrics import WordErrorRate
+
+
+class SpeakerAwareTranscriptionConfig(PipelineConfig):
+    def __init__(
+        self,
+        asr: Optional[m.SpeechRecognitionModel] = None,
+        segmentation: Optional[m.SegmentationModel] = None,
+        embedding: Optional[m.EmbeddingModel] = None,
+        duration: Optional[float] = None,
+        asr_duration: float = 3,
+        step: float = 0.5,
+        latency: Optional[Union[float, Literal["max", "min"]]] = None,
+        tau_active: float = 0.5,
+        rho_update: float = 0.3,
+        delta_new: float = 1,
+        language: Optional[Text] = None,
+        beam_size: Optional[int] = None,
+        gamma: float = 3,
+        beta: float = 10,
+        max_speakers: int = 20,
+        merge_collar: float = 0.05,
+        diarization_device: Optional[torch.device] = None,
+        asr_device: Optional[torch.device] = None,
+        **kwargs,
+    ):
+        # Default segmentation model is pyannote/segmentation
+        self.segmentation = segmentation
+        if self.segmentation is None:
+            self.segmentation = m.SegmentationModel.from_pyannote("pyannote/segmentation")
+
+        self._duration = duration
+        self._sample_rate: Optional[int] = None
+
+        # Default embedding model is pyannote/embedding
+        self.embedding = embedding
+        if self.embedding is None:
+            self.embedding = m.EmbeddingModel.from_pyannote("pyannote/embedding")
+
+        # Latency defaults to the step duration
+        self._step = step
+        self._latency = latency
+        if self._latency is None or self._latency == "min":
+            self._latency = self._step
+        elif self._latency == "max":
+            self._latency = self.duration
+
+        self.tau_active = tau_active
+        self.rho_update = rho_update
+        self.delta_new = delta_new
+        self.gamma = gamma
+        self.beta = beta
+        self.max_speakers = max_speakers
+        self.merge_collar = merge_collar
+        self.asr_duration = asr_duration
+
+        self.diarization_device = diarization_device
+        if self.diarization_device is None:
+            self.diarization_device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+
+        self.language = language
+        self.beam_size = beam_size
+
+        self.asr_device = asr_device
+        if self.asr_device is None:
+            self.asr_device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+
+        # Default ASR model is Whisper small (244M parameters)
+        self.asr = asr
+        if self.asr is None:
+            self.asr = m.SpeechRecognitionModel.from_whisper("small")
+        self.asr.set_language(self.language)
+        self.asr.set_beam_size(self.beam_size)
+
+    def to_diarization_config(self) -> SpeakerDiarizationConfig:
+        return SpeakerDiarizationConfig(
+            segmentation=self.segmentation,
+            embedding=self.embedding,
+            duration=self.duration,
+            step=self.step,
+            latency=self.latency,
+            tau_active=self.tau_active,
[…]
