### 2023-04-26 c92c320
Playing with clustering constraints and aggregation
