### 2023-04-26 c7bbcc4
Refactor SpeakerAwareTranscription
