+++ src/diart/pipelines/speaker_transcription.py
@@ -209,13 +209,8 @@ class SpeakerAwareTranscription(Pipeline):
     def suggest_writer(self, uri: Text, output_dir: Union[Text, Path]) -> Observer:
         return sinks.TextWriter(Path(output_dir) / f"{uri}.txt")
 
-    def __call__(
-        self,
-        waveforms: Sequence[SlidingWindowFeature],
-    ) -> Sequence[Text]:
-        # Compute diarization output
-        diarization_output = self.diarization(waveforms)
-
+    def _update_buffers(self, diarization_output: Sequence[Tuple[Annotation, SlidingWindowFeature]]):
+        # Separate diarization and aligned audio chunks
         first_chunk = diarization_output[0][1]
         output_start = first_chunk.extent.start
         resolution = first_chunk.sliding_window.duration
@@ -239,78 +234,109 @@ class SpeakerAwareTranscription(Pipeline):
             chunk_data.insert(0, self.audio_buffer.data)
             self.audio_buffer = SlidingWindowFeature(
                 np.concatenate(chunk_data, axis=0),
-                self.audio_buffer.sliding_window
+                self.audio_buffer.sliding_window,
             )
 
-        # Extract audio to transcribe from the buffer
-        asr_duration = self.config.asr_duration
+    def _extract_asr_inputs(self) -> Tuple[List[SlidingWindowFeature], List[Annotation]]:
+        chunk_duration = self.config.asr_duration
         buffer_duration = self.audio_buffer.extent.duration
-        asr_batch_size = int(buffer_duration / asr_duration)
-
-        if asr_batch_size == 0:
-            return ["" for _ in waveforms]
-
+        batch_size = int(buffer_duration / chunk_duration)
         buffer_start = self.audio_buffer.extent.start
+        resolution = self.audio_buffer.sliding_window.duration
+
+        # Extract audio chunks with their diarization
         asr_inputs, input_dia, last_end_time = [], [], None
-        for i in range(asr_batch_size):
-            start = buffer_start + i * asr_duration
-            last_end_time = start + asr_duration
+        for i in range(batch_size):
+            start = buffer_start + i * chunk_duration
+            last_end_time = start + chunk_duration
             region = Segment(start, last_end_time)
-            chunk = self.audio_buffer.crop(region, fixed=asr_duration)
+            chunk = self.audio_buffer.crop(region, fixed=chunk_duration)
             window = SlidingWindow(resolution, resolution, start)
             asr_inputs.append(SlidingWindowFeature(chunk, window))
             input_dia.append(self.dia_buffer.crop(region))
 
-        # Create ASR batch, shape (batch, samples, channels)
-        batch = torch.stack([torch.from_numpy(w.data) for w in asr_inputs])
+        # Remove extracted chunks from buffers
+        if asr_inputs:
+            new_buffer_bounds = Segment(last_end_time, self.audio_buffer.extent.end)
+            new_buffer = self.audio_buffer.crop(new_buffer_bounds, fixed=new_buffer_bounds.duration)
+            window = SlidingWindow(resolution, resolution, last_end_time)
+            self.audio_buffer = SlidingWindowFeature(new_buffer, window)
+            self.dia_buffer = self.dia_buffer.extrude(Segment(0, last_end_time))
 
-        # Remove transcribed chunks from buffer
-        new_buffer_bounds = Segment(last_end_time, self.audio_buffer.extent.end)
-        new_buffer = self.audio_buffer.crop(new_buffer_bounds, fixed=new_buffer_bounds.duration)
-        window = SlidingWindow(resolution, resolution, last_end_time)
-        self.audio_buffer = SlidingWindowFeature(new_buffer, window)
[…]
