+++ src/diart/console/benchmark.py
@@ -24,6 +24,8 @@ def run():
                         help="Optional. Directory with RTTM files CONVERSATION.rttm. Names must match audio files")
     parser.add_argument("--duration", default=5, type=float,
                         help=f"Duration of the sliding window (in seconds). Default value depends on the pipeline")
+    parser.add_argument("--asr-duration", default=3, type=float,
+                        help=f"Duration of the transcription window (in seconds). Defaults to 3")
     parser.add_argument("--step", default=0.5, type=float, help=f"{argdoc.STEP}. Defaults to 0.5")
     parser.add_argument("--latency", default=0.5, type=float, help=f"{argdoc.LATENCY}. Defaults to 0.5")
     parser.add_argument("--tau", default=0.5, type=float, help=f"{argdoc.TAU}. Defaults to 0.5")
