### 2023-05-21 d3f008e
Add automatic sample rate detection in MicrophoneAudioSource. Fix resampling crash.
