### 2023-06-23 juanmc2005/diart#156 @sneakers-the-rat
[joss] Tests?
Hey! Sorry for the long delay. just started a new job and things have been hectic.

Orienting myself to the package, and I can't seem to find any tests? JOSS doesn't have a minimum test coverage requirement or anything like that, but it does ask that essential functionality is tested, and I have interpreted that in the past as meaning unit tests that cover most of the major operations described in the docs and the paper.

I'll tag all my issues from the review with [joss] and also link back to the review so we can track them from there

Part of:
https://github.com/openjournals/joss-reviews/issues/5266
