### 2023-06-27 juanmc2005/diart#156 @juanmc2005
Hi @sneakers-the-rat ! Thanks for your feedback! I've been wanting to add unit tests for some time (#94).
I will prioritize it so they are implemented for the main features of the library.

Otherwise, the core operations of the library can also be verified using the CLI interface: `diart.stream` for the microphone and individual files, and `diart.benchmark` to reproduce the performance that we claim (e.g. on the AMI dataset).
