### 2023-06-28 juanmc2005/diart#156 @sneakers-the-rat
I'd be happy to make a pull request adding some tests as examples, if you haven't written them before :)
