### 2023-06-29 juanmc2005/diart#157 @sneakers-the-rat
[joss] Inconsistent naming causes cli arguments to be ignored
Trying to write a simple example test using the config and found a bug

Argument in argparse is named `tau`

https://github.com/juanmc2005/diart/blob/858cd1a6a8024d74aa91f70fd5352dc37128230d/src/diart/console/benchmark.py#L21

passed to `from_dict`, which looks for `tau_active` instead, replacing whatever i feed to it with the default:

https://github.com/juanmc2005/diart/blob/858cd1a6a8024d74aa91f70fd5352dc37128230d/src/diart/blocks/config.py#L110

Same is true of `rho` -> `rho_update` and `delta` -> `delta_new`.

Some additional things i noticed in these modules that aren't bugs per se but just some tips:

- The initialization logic is written twice, once in `__init__` and once in `from_dict` without clear purpose. this makes for very fragile and bug-prone code. I'm not sure what the `from_dict` method is supposed to do differently than just calling the class like `PipelineConfig(**args)` - my expectation as a user would be that they would be identical. In fact this seems to be the cause of the bug at hand - the `**kwargs` in `__init__` is superfluous since they are ignored, but if it wasn't there and I called `Pipelineconfig(**args)` then I would have gotten an exception from passing an unexpected parameter.
- The default values are actually defined three times - another time in the argparser itself. Definitely recommend having a single place these are defined.
- you probably want `@classmethod` here: https://github.com/juanmc2005/diart/blob/858cd1a6a8024d74aa91f70fd5352dc37128230d/src/diart/blocks/config.py#L95
- try using the `abc` module for abstract classes like eg. - https://github.com/juanmc2005/diart/blob/858cd1a6a8024d74aa91f70fd5352dc37128230d/src/diart/blocks/config.py#L12
- this signature should probably be `data: dict` since the method is named "from_dict" - https://github.com/juanmc2005/diart/blob/858cd1a6a8024d74aa91f70fd5352dc37128230d/src/diart/blocks/config.py#L30
- `dict`s already have a get method, eg. `{'a': 'dict'}.get('a', None)` that works the same as `utils.get`
- the `__init__` methods are very heavy for a config class - instantiating two models. If this is just a config file you should probably just store the string values and instantiate the models elsewhere.

part of https://github.com/openjournals/joss-reviews/issues/5266
