### 2023-06-29 juanmc2005/diart#158 @sneakers-the-rat
[joss] Outdated torchaudio can't find newer ffmpeg, conda environment.yml file missing to specify version
Trying to run the benchmark, and I got an import error from torchaudio:

```
_______________________________________________ ERROR collecting tests/test_config.py ________________________________________________
ImportError while importing test module 'diart_fork/tests/test_config.py'.
Traceback:
venv/lib/python3.9/site-packages/torchaudio/_extension.py:71: in _init_ffmpeg
    _load_lib("libtorchaudio_ffmpeg")
venv/lib/python3.9/site-packages/torchaudio/_extension.py:52: in _load_lib
    torch.ops.load_library(path)
venv/lib/python3.9/site-packages/torch/_ops.py:573: in load_library
    ctypes.CDLL(path)
../../.pyenv/versions/3.9.1/lib/python3.9/ctypes/__init__.py:374: in __init__
    self._handle = _dlopen(self._name, mode)
E   OSError: dlopen(diart_fork/venv/lib/python3.9/site-packages/torchaudio/lib/libtorchaudio_ffmpeg.so, 0x0006): Library not loaded: '@rpath/libavdevice.58.dylib'
E     Referenced from: 'diart_fork/venv/lib/python3.9/site-packages/torchaudio/lib/libtorchaudio_ffmpeg.so'
E     Reason: tried: '/usr/local/lib/libavdevice.58.dylib' (no such file), '/usr/lib/libavdevice.58.dylib' (no such file)

../../.pyenv/versions/3.9.1/lib/python3.9/importlib/__init__.py:127: in import_module
tests/test_config.py:7: in <module>
    from diart.console.benchmark import parse_args
src/diart/console/benchmark.py:9: in <module>
    from diart.inference import Benchmark, Parallelize
src/diart/inference.py:8: in <module>
    import diart.sources as src
src/diart/sources.py:11: in <module>
    from torchaudio.io import StreamReader
venv/lib/python3.9/site-packages/torchaudio/io/__init__.py:21: in __getattr__
    torchaudio._extension._init_ffmpeg()
venv/lib/python3.9/site-packages/torchaudio/_extension.py:73: in _init_ffmpeg
    raise ImportError("FFmpeg libraries are not found. Please install FFmpeg.") from err
E   ImportError: FFmpeg libraries are not found. Please install FFmpeg.
```

This apparently was also raised on stack overflow: https://stackoverflow.com/questions/76155851/diart-torchaudio-on-windows-x64-results-in-torchaudio-error-importerror-ffmp

I am not going to debug it right now, but this would have been immediately discovered in  CI - the installation instructions are to manually create and populate a conda environment, and if the problem is indeed that the old version of torchaudio can't handle newer ffmpegs, then that would have been caught trying to install that way in CI because conda-forge's version is 6.0.0 at the moment: https://anaconda.org/conda-forge/ffmpeg

If the package depends on a conda environment, then I would expect an `environment.yml` file to be included with the package that specifies the appropriate dependencies.

Part of: https://github.com/openjournals/joss-reviews/issues/5266
