### 2023-06-29 juanmc2005/diart#156 @sneakers-the-rat
Here's an example test for the config:
https://github.com/sneakers-the-rat/diart/commit/b1a0ccaa35f8b36aa30f978a4bcb16db69652a42

after installing with `[tests]`, one can run `pytest` from the repo root to run the tests. Gave an example of using test fixtures for random value init to test that the inputs arent being ignored and defaults substituted
