### 2023-06-30 juanmc2005/diart#157 @juanmc2005
Hi @sneakers-the-rat,

`tau_active` and `tau` are aliases, you can see in the following lines that it checks for `tau_active` first, if that fails it looks for `tau` defaulting to 0.6 (see [here](https://github.com/juanmc2005/diart/blob/858cd1a6a8024d74aa91f70fd5352dc37128230d/src/diart/blocks/config.py#L112)). The same applies for the other arguments you mention.
Notice that the cli argument `max-speakers` is converted to `max_speakers` automatically by argparse.

I do agree on the comments about the initialization code and the repetitiveness. Do you have any suggestions for a pythonic way of addressing this?
