### 2023-06-30 juanmc2005/diart#158 @sneakers-the-rat
I think the fix is as simple as making a conda environment file, and I think torchaudio has an updated version?

To prevent future bugs like this, a very simple CI (eg. github actions) workflow that installs the package and runs the benchmark would at least ensure that the package runs
