### 2023-06-30 juanmc2005/diart#157 @sneakers-the-rat
Ah, my bad, something was causing my test to fail and it was late so i must not have read carefully.

Yes! I am unsure what the `from_dict` method is supposed to do differently than the `__init__` method - i test that they are identical here: https://github.com/sneakers-the-rat/diart/blob/b1a0ccaa35f8b36aa30f978a4bcb16db69652a42/tests/test_config.py#L35

usually when you have a `from_x` method it's because the thing you're instantiating from is different in form than how you usually instantiate it - eg. a `from_json` method might take a path to a `.json` file with instantiation arguments. In this case, instantiating like `MyClass(**{'arg1':'val1', 'arg2':'val2'})` is the same as `MyClass(arg1='val1', arg2='val2')`

So my suggestion would be to remove the `from_dict` method and use the same variable names everywhere.

I don't mean to harp on you about a minor thing, I am more focused on showing you how to make the code more testable by removing special cases, minimizing where things can go wrong, etc. I won't be able to review the entire package, so this is just one example of a strategy for avoiding bugs.
