### 2023-07-12 juanmc2005/diart#158 @zaouk
@sneakers-the-rat on which platform did you try to run the benchmark? Is it windows too (as per the stackoverflow link that you shared)? If so, did you try to use diart on WSL instead?

I am using linux, and I can't seem to be able to reproduce the error that you mentioned. I created a new conda environment and followed the steps for installation (first running `conda install portaudio pysoundfile ffmpeg -c conda-forge` and then `pip install diart`) and everything works properly.

FWIW, I exported the environment I have into this environment yml file:

```
name: diart
channels:
  - conda-forge
  - defaults
dependencies:
  - _libgcc_mutex=0.1=conda_forge
  - _openmp_mutex=4.5=2_kmp_llvm
  - alsa-lib=1.2.9=hd590300_0
  - bzip2=1.0.8=h7f98852_4
  - ca-certificates=2023.5.7=hbcca054_0
  - cffi=1.15.1=py38h4a40e3a_3
  - ffmpeg=4.2.2=h20bf706_0
  - freetype=2.10.4=h0708190_1
  - gettext=0.21.1=h27087fc_0
  - gmp=6.2.1=h58526e2_0
  - gnutls=3.6.13=h85f3911_1
  - lame=3.100=h166bdaf_1003
  - ld_impl_linux-64=2.38=h1181459_1
  - libblas=3.9.0=17_linux64_openblas
  - libcblas=3.9.0=17_linux64_openblas
  - libffi=3.4.4=h6a678d5_0
  - libflac=1.4.3=h59595ed_0
  - libgcc-ng=13.1.0=he5830b7_0
  - libgfortran-ng=13.1.0=h69a702a_0
  - libgfortran5=13.1.0=h15d22d2_0
  - liblapack=3.9.0=17_linux64_openblas
  - libogg=1.3.4=h7f98852_1
  - libopenblas=0.3.23=pthreads_h80387f5_0
  - libopus=1.3.1=h7f98852_1
  - libpng=1.6.39=h5eee18b_0
  - libsndfile=1.2.0=hb75c966_0
  - libstdcxx-ng=13.1.0=hfd8a6a1_0
  - libvorbis=1.3.7=h9c3ff4c_0
  - libvpx=1.7.0=h439df22_0
  - llvm-openmp=14.0.6=h9e868ea_0
  - mpg123=1.31.3=hcb278e6_0
  - ncurses=6.4=h6a678d5_0
  - nettle=3.6=he412f7d_0
  - numpy=1.24.4=py38h59b608b_0
  - openh264=2.1.1=h4ff587b_0
  - openssl=3.1.1=hd590300_1
  - pip=23.1.2=py38h06a4308_0
  - portaudio=19.6.0=h0e77e87_8
  - pycparser=2.21=pyhd8ed1ab_0
  - pysoundfile=0.12.1=pyhd8ed1ab_0
  - python=3.8.17=h955ad1f_0
  - python_abi=3.8=2_cp38
  - readline=8.2=h5eee18b_0
  - setuptools=67.8.0=py38h06a4308_0
  - sqlite=3.41.2=h5eee18b_0
  - tk=8.6.12=h1ccaba5_0
  - wheel=0.38.4=py38h06a4308_0
  - x264=1!157.20191217=h7b6447c_0
  - xz=5.4.2=h5eee18b_0
  - zlib=1.2.13=h5eee18b_0
  - pip:
      - absl-py==1.4.0
      - aiohttp==3.8.4
      - aiosignal==1.3.1
      - alembic==1.11.1
      - antlr4-python3-runtime==4.9.3
      - asteroid-filterbanks==0.4.0
      - async-timeout==4.0.2
      - attrs==23.1.0
      - audioread==3.0.0
      - backports-cached-property==1.0.2
      - cachetools==5.3.1
      - certifi==2023.5.7
      - charset-normalizer==3.2.0
      - click==8.1.4
      - cmaes==0.9.1
      - colorama==0.4.6
      - colorlog==6.7.0
      - contourpy==1.1.0
      - cycler==0.11.0
      - decorator==5.1.1
      - diart==0.7.0
      - docopt==0.6.2
      - einops==0.3.2
      - filelock==3.12.2
      - fonttools==4.41.0
      - frozenlist==1.4.0
      - fsspec==2023.6.0
      - google-auth==2.22.0
      - google-auth-oauthlib==1.0.0
      - greenlet==2.0.2
      - grpcio==1.56.0
      - hmmlearn==0.2.8
      - huggingface-hub==0.16.4
      - hyperpyyaml==1.2.1
      - idna==3.4
      - importlib-metadata==6.8.0
      - importlib-resources==6.0.0
      - joblib==1.3.1
      - julius==0.2.7
      - kiwisolver==1.4.4
      - librosa==0.9.2
      - llvmlite==0.40.1
      - mako==1.2.4
      - markdown==3.4.3
      - markdown-it-py==3.0.0
      - markupsafe==2.1.3
      - matplotlib==3.7.2
      - mdurl==0.1.2
      - mpmath==1.3.0
      - multidict==6.0.4
      - networkx==2.8.8
      - numba==0.57.1
      - nvidia-cublas-cu11==11.10.3.66
      - nvidia-cuda-nvrtc-cu11==11.7.99
      - nvidia-cuda-runtime-cu11==11.7.99
      - nvidia-cudnn-cu11==8.5.0.96
      - oauthlib==3.2.2
      - omegaconf==2.3.0
      - optuna==3.2.0
      - packaging==23.1
      - pandas==2.0.3
      - pillow==10.0.0
      - platformdirs==3.8.1
      - pooch==1.7.0
      - primepy==1.3
      - protobuf==3.20.1
      - pyannote-audio==2.1.1
      - pyannote-core==4.5
      - pyannote-database==4.1.3
      - pyannote-metrics==3.2.1
      - pyannote-pipeline==2.3
      - pyasn1==0.5.0
      - pyasn1-modules==0.3.0
      - pydeprecate==0.3.2
      - pygments==2.15.1
      - pyparsing==3.0.9
      - python-dateutil==2.8.2
      - pytorch-lightning==1.6.5
      - pytorch-metric-learning==1.7.3
      - pytz==2023.3
      - pyyaml==6.0
      - requests==2.31.0
      - requests-oauthlib==1.3.1
      - resampy==0.4.2
      - rich==13.4.2
      - rsa==4.9
      - ruamel-yaml==0.17.28
      - ruamel-yaml-clib==0.2.7
      - rx==3.2.0
      - scikit-learn==1.3.0
      - scipy==1.10.1
      - semver==2.13.0
      - sentencepiece==0.1.99
      - shellingham==1.5.0.post1
      - simplejson==3.19.1
      - singledispatchmethod==1.0
      - six==1.16.0
      - sortedcontainers==2.4.0
      - sounddevice==0.4.6
      - soundfile==0.10.3.post1
      - speechbrain==0.5.14
      - sqlalchemy==2.0.18
      - sympy==1.12
      - tabulate==0.9.0
      - tensorboard==2.13.0
      - tensorboard-data-server==0.7.1
      - threadpoolctl==3.1.0
      - torch==1.13.1
      - torch-audiomentations==0.11.0
      - torch-pitch-shift==1.2.4
      - torchaudio==0.13.1
      - torchmetrics==0.11.4
      - torchvision==0.14.1
      - tqdm==4.65.0
      - typer==0.9.0
      - typing-extensions==4.7.1
      - tzdata==2023.3
      - urllib3==1.26.16
      - websocket-client==1.6.1
      - websocket-server==0.6.4
      - werkzeug==2.3.6
      - yarl==1.9.2
      - zipp==3.16.0
```
