### 2023-07-13 juanmc2005/diart#158 @juanmc2005
I confirm this can be solved with `ffmpeg<4.4`.
A temporary fix would be to change the installation instructions:

```diff
- conda install portaudio pysoundfile ffmpeg -c conda-forge
+ conda install portaudio=19.6.0 pysoundfile=0.12.1 ffmpeg=4.3 -c conda-forge
```
