### 2023-07-19 juanmc2005/diart#158 @juanmc2005
I agree, the `environment.yml` file is the correct way to do this. I'm just pointing out a solution for people coming to this issue with the ffmpeg error until I have the time to make the required changes.
