### 2023-10-04 98bbb4f
Replace NotImplementedError with ABC/abstractmethod
