### 2023-10-04 5f1816d
Replace NotImplementedError with ABC/abstractmethod
