### 2023-10-04 juanmc2005/diart#174 @sneakers-the-rat
[joss][ux][minor] Catch `KeyboardInterrupt` in `diart.stream`
The `diart.stream` method seems to work well! when I exit with ctrl+c though it just crashes out - would be nice to add a `try: ... except KeyboardInterrupt:` here: https://github.com/juanmc2005/diart/blob/858cd1a6a8024d74aa91f70fd5352dc37128230d/src/diart/console/stream.py#L63

related to: https://github.com/openjournals/joss-reviews/issues/5266
