### 2023-10-04 juanmc2005/diart#175 @sneakers-the-rat
[joss] Need locations of benchmarking data!
Trying to find the data to run the benchmarks, and I can't find all the source data:

- [x] **VoxConverse** - found it!
- [ ] **AMI** - found the download page https://groups.inf.ed.ac.uk/ami/download/ but I can't tell what files I need for the benchmark
- [ ] **DIHARD III* - can't find download
- [ ] **DIHARD II* - can't find download

this should all be in the docs!

part of https://github.com/openjournals/joss-reviews/issues/5266
