### 2023-10-04 juanmc2005/diart#176 @sneakers-the-rat
[joss][docs] - Collected Docs suggestions
Just going to gather all my thoughts on the docs in this issue rather than raising separate issues

- Following JOSS's checklist, you're supposed to have some statement of need in the docs. Would highly recommend at least adding a description of what the package does to the top of the README, y'all just jump right in to installation :)
- Relatedly, would be good to explain how this package is different than others - some brief description or nod to what you do differently! give yourself some credit for having good ideas :)
- Would be good to add some descriptions of what the data structures that are created at various points - eg. in the hyperparameter tuning I'm not sure what is created or how to access it.
- Not everything needs to go on readthedocs, but i'm not sure why this package isn't since there is a reasonable amount of complexity and some nice moments in the source that would be nice to have further explanation of. At least an overview of the package structure would be nice (as simple (simple as a good thing!) as it is). That and some of the "core" classes should be explained like the segmentation/diarization classes. There is some explanation in the paper within the repo, but more implementation details would be welcome.

Part of: https://github.com/openjournals/joss-reviews/issues/5266
