### 2023-10-04 #5266 @sneakers-the-rat
OK! finished my review, hopefully my general review at the bottom answers any questions, and open issues and pull requests are suggestions! I'll leave any further decision up to the editor here, i'm a "pass" on this one since i view my role as a reviewer as helping the authors improve the package, not determining whether it should be in or out of an exclusive club <3. Lovely package, thanks for the opportunity to look at it, and again sorry it took so long. Cheers to the authors!
