### 2023-10-04 juanmc2005/diart#175 @sneakers-the-rat
Forgot to say - I also couldnt tell easily from the text of the repo paper which samples I should use from voxconverse, might be nice to have a link to a source and a little script for loading them - making reproducing like 1 step if you have a spare half hour or so
