### 2023-10-06 juanmc2005/diart#176 @juanmc2005
@sneakers-the-rat not sure I understand the last two points. Could you expand on those?
