### 2023-10-06 juanmc2005/diart#175 @juanmc2005
@sneakers-the-rat good idea! However, the two DIHARD datasets are private. One has to ask for permission to use them.
For AMI we used Mix-Headset (downloadable [here](https://github.com/pyannote/AMI-diarization-setup/tree/main/pyannote))
