### 2023-10-06 dff9e80
[joss] Add Conda environment (#172)
* Conda environment!

* Update README with env instructions

---------

Co-authored-by: Juan Coria <juanmc2005@hotmail.com>
