### 2023-10-06 juanmc2005/diart#158 @juanmc2005
Conda `environment.yml` added to `develop`. It will appear in the next release
