### 2023-10-11 fa4eecd
Replace NotImplementedError with ABC/abstractmethod (#179)
Co-authored-by: sneakers-the-rat <JLSaunders987@gmail.com>
