### 2023-10-11 d43cacf
Replace NotImplementedError with ABC/abstractmethod (#179)
Co-authored-by: sneakers-the-rat <JLSaunders987@gmail.com>
