### 2023-10-11 8299b70
Fix sample rate issues (#153)
* Add automatic sample rate detection in MicrophoneAudioSource. Fix resampling crash.

* Replace block_size by block_duration in audio source constructors
