### 2023-10-11 410ab89
Fix sample rate issues (#153)
* Add automatic sample rate detection in MicrophoneAudioSource. Fix resampling crash.

* Replace block_size by block_duration in audio source constructors
