+++ setup.cfg
@@ -30,7 +30,7 @@ install_requires=
     pandas>=1.4.2
     torch>=1.12.1
     torchvision>=0.14.0
-    torchaudio>=0.12.1,<1.0
+    torchaudio>=2.0.2
     pyannote.audio>=2.1.1
     pyannote.core>=4.5
     pyannote.database>=4.1.1
