+++ src/diart/console/tune.py
@@ -11,34 +11,95 @@ from optuna.samplers import TPESampler
 
 def run():
     parser = argparse.ArgumentParser()
-    parser.add_argument("root", type=str, help="Directory with audio files CONVERSATION.(wav|flac|m4a|...)")
-    parser.add_argument("--reference", required=True, type=str,
-                        help="Directory with RTTM files CONVERSATION.rttm. Names must match audio files")
-    parser.add_argument("--pipeline", default="SpeakerDiarization", type=str,
-                        help="Class of the pipeline to optimize. Defaults to 'SpeakerDiarization'")
-    parser.add_argument("--segmentation", default="pyannote/segmentation", type=str,
-                        help=f"{argdoc.SEGMENTATION}. Defaults to pyannote/segmentation")
-    parser.add_argument("--embedding", default="pyannote/embedding", type=str,
-                        help=f"{argdoc.EMBEDDING}. Defaults to pyannote/embedding")
-    parser.add_argument("--step", default=0.5, type=float, help=f"{argdoc.STEP}. Defaults to 0.5")
-    parser.add_argument("--latency", default=0.5, type=float, help=f"{argdoc.LATENCY}. Defaults to 0.5")
-    parser.add_argument("--tau", default=0.5, type=float, help=f"{argdoc.TAU}. Defaults to 0.5")
-    parser.add_argument("--rho", default=0.3, type=float, help=f"{argdoc.RHO}. Defaults to 0.3")
-    parser.add_argument("--delta", default=1, type=float, help=f"{argdoc.DELTA}. Defaults to 1")
-    parser.add_argument("--gamma", default=3, type=float, help=f"{argdoc.GAMMA}. Defaults to 3")
-    parser.add_argument("--beta", default=10, type=float, help=f"{argdoc.BETA}. Defaults to 10")
-    parser.add_argument("--max-speakers", default=20, type=int, help=f"{argdoc.MAX_SPEAKERS}. Defaults to 20")
-    parser.add_argument("--batch-size", default=32, type=int, help=f"{argdoc.BATCH_SIZE}. Defaults to 32")
-    parser.add_argument("--cpu", dest="cpu", action="store_true",
-                        help=f"{argdoc.CPU}. Defaults to GPU if available, CPU otherwise")
-    parser.add_argument("--hparams", nargs="+", default=("tau_active", "rho_update", "delta_new"),
-                        help="Hyper-parameters to optimize. Must match names in `PipelineConfig`. Defaults to tau_active, rho_update and delta_new")
-    parser.add_argument("--num-iter", default=100, type=int, help="Number of optimization trials")
-    parser.add_argument("--storage", type=str,
-                        help="Optuna storage string. If provided, continue a previous study instead of creating one. The database name must match the study name")
+    parser.add_argument(
+        "root",
+        type=str,
+        help="Directory with audio files CONVERSATION.(wav|flac|m4a|...)",
+    )
+    parser.add_argument(
+        "--reference",
+        required=True,
+        type=str,
+        help="Directory with RTTM files CONVERSATION.rttm. Names must match audio files",
+    )
+    parser.add_argument(
+        "--pipeline",
+        default="SpeakerDiarization",
+        type=str,
+        help="Class of the pipeline to optimize. Defaults to 'SpeakerDiarization'",
+    )
+    parser.add_argument(
+        "--segmentation",
+        default="pyannote/segmentation",
+        type=str,
+        help=f"{argdoc.SEGMENTATION}. Defaults to pyannote/segmentation",
+    )
+    parser.add_argument(
+        "--embedding",
+        default="pyannote/embedding",
[…]
