+++ src/diart/console/client.py
@@ -21,9 +21,7 @@ def send_audio(ws: WebSocket, source: Text, step: float, sample_rate: int):
         audio_source = src.MicrophoneAudioSource(step, device)
 
     # Encode audio, then send through websocket
-    audio_source.stream.pipe(
-        ops.map(utils.encode_audio)
-    ).subscribe_(ws.send)
+    audio_source.stream.pipe(ops.map(utils.encode_audio)).subscribe_(ws.send)
 
     # Start reading audio
     audio_source.read()
@@ -40,18 +38,37 @@ def receive_audio(ws: WebSocket, output: Optional[Path]):
 
 def run():
     parser = argparse.ArgumentParser()
-    parser.add_argument("source", type=str, help="Path to an audio file | 'microphone' | 'microphone:<DEVICE_ID>'")
+    parser.add_argument(
+        "source",
+        type=str,
+        help="Path to an audio file | 'microphone' | 'microphone:<DEVICE_ID>'",
+    )
     parser.add_argument("--host", required=True, type=str, help="Server host")
     parser.add_argument("--port", required=True, type=int, help="Server port")
-    parser.add_argument("--step", default=0.5, type=float, help=f"{argdoc.STEP}. Defaults to 0.5")
-    parser.add_argument("-sr", "--sample-rate", default=16000, type=int, help=f"{argdoc.SAMPLE_RATE}. Defaults to 16000")
-    parser.add_argument("-o", "--output-file", type=Path, help="Output RTTM file. Defaults to no writing")
+    parser.add_argument(
+        "--step", default=0.5, type=float, help=f"{argdoc.STEP}. Defaults to 0.5"
+    )
+    parser.add_argument(
+        "-sr",
+        "--sample-rate",
+        default=16000,
+        type=int,
+        help=f"{argdoc.SAMPLE_RATE}. Defaults to 16000",
+    )
+    parser.add_argument(
+        "-o",
+        "--output-file",
+        type=Path,
+        help="Output RTTM file. Defaults to no writing",
+    )
     args = parser.parse_args()
 
     # Run websocket client
     ws = WebSocket()
     ws.connect(f"ws://{args.host}:{args.port}")
-    sender = Thread(target=send_audio, args=[ws, args.source, args.step, args.sample_rate])
+    sender = Thread(
+        target=send_audio, args=[ws, args.source, args.step, args.sample_rate]
+    )
     receiver = Thread(target=receive_audio, args=[ws, args.output_file])
     sender.start()
     receiver.start()

[…]
