+++ src/diart/operators.py
@@ -25,20 +25,24 @@ class AudioBufferState:
     def has_samples(num_samples: int):
         def call_fn(state) -> bool:
             return state.chunk is not None and state.chunk.shape[1] == num_samples
+
         return call_fn
 
     @staticmethod
     def to_sliding_window(sample_rate: int):
         def call_fn(state) -> SlidingWindowFeature:
-            resolution = SlidingWindow(start=state.start_time, duration=1. / sample_rate, step=1. / sample_rate)
+            resolution = SlidingWindow(
+                start=state.start_time,
+                duration=1.0 / sample_rate,
+                step=1.0 / sample_rate,
+            )
             return SlidingWindowFeature(state.chunk.T, resolution)
+
         return call_fn
 
 
 def rearrange_audio_stream(
-    duration: float = 5,
-    step: float = 0.5,
-    sample_rate: int = 16000
+    duration: float = 5, step: float = 0.5, sample_rate: int = 16000
 ) -> Operator:
     chunk_samples = int(round(sample_rate * duration))
     step_samples = int(round(sample_rate * step))
@@ -49,11 +53,17 @@ def rearrange_audio_stream(
     def accumulate(state: AudioBufferState, value: np.ndarray):
         # State contains the last emitted chunk, the current step buffer and the last starting time
         if value.ndim != 2 or value.shape[0] != 1:
-            raise ValueError(f"Waveform must have shape (1, samples) but {value.shape} was found")
+            raise ValueError(
+                f"Waveform must have shape (1, samples) but {value.shape} was found"
+            )
         start_time = state.start_time
 
         # Add new samples to the buffer
-        buffer = value if state.buffer is None else np.concatenate([state.buffer, value], axis=1)
+        buffer = (
+            value
+            if state.buffer is None
+            else np.concatenate([state.buffer, value], axis=1)
+        )
 
         # Check for buffer overflow
         if buffer.shape[1] >= step_samples:
@@ -86,7 +96,7 @@ def rearrange_audio_stream(
         ops.filter(AudioBufferState.has_samples(chunk_samples)),
         ops.filter(lambda state: state.changed),
         # Transform state into a SlidingWindowFeature containing the new chunk
-        ops.map(AudioBufferState.to_sliding_window(sample_rate))
+        ops.map(AudioBufferState.to_sliding_window(sample_rate)),
     )
 
 
@@ -96,6 +106,7 @@ def buffer_slide(n: int):
         if len(new_state) > n:
             return new_state[1:]
         return new_state
+
     return rx.pipe(ops.scan(accumulate, []))
 
 
@@ -117,17 +128,19 @@ class OutputAccumulationState:
     next_sample: Optional[int]
 
     @staticmethod
-    def initial() -> 'OutputAccumulationState':
+    def initial() -> "OutputAccumulationState":
         return OutputAccumulationState(None, None, 0, 0)
 
     @property
     def cropped_waveform(self) -> SlidingWindowFeature:
         return SlidingWindowFeature(
-            self.waveform[:self.next_sample],
+            self.waveform[: self.next_sample],
             self.waveform.sliding_window,
         )
 
-    def to_tuple(self) -> Tuple[Optional[Annotation], Optional[SlidingWindowFeature], float]:
+    def to_tuple(
+        self,
+    ) -> Tuple[Optional[Annotation], Optional[SlidingWindowFeature], float]:
         return self.annotation, self.cropped_waveform, self.real_time
 
 
@@ -153,9 +166,10 @@ def accumulate_output(
     -------
     A reactive x operator implementing this behavior.
     """
+
     def accumulate(
         state: OutputAccumulationState,
[…]
