### 2023-10-11 45f8ad9
Catch keyboard interrupt in diart.stream (#183)
