### 2023-10-11 juanmc2005/diart#174 @juanmc2005
Implemented in #183
