### 2023-10-11 juanmc2005/diart#176 @sneakers-the-rat
Yes! So for 3, eg [here](https://github.com/juanmc2005/diart#distributed-optimization) in the optimization part, I don't know what the database tables that would create are - in general when you have some sort of data structure that's produced or consumed by a package it's good to give some kind of abbreviated example except when it's a standard (eg. it seems like .rttm is a standard so you wouldn't necessarily need to provide an example of it, but even providing a 5-line `head` of one would be useful).

For 4, there are decent "how to use this" docs in the README, but I couldn't find any docs about "here is how the package is structured" or how individual classes/functions worked. Some of the classes/functions already have docstrings, so it would be pretty nice to make that work you've done doing internal API-level documentation visible by using something like [sphinx](https://docs.readthedocs.io/en/stable/intro/getting-started-with-sphinx.html) and [autodoc](https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html) to create a documentation website.

so eg here's an old package of mine's documentation - https://docs.auto-pi-lot.com/en/latest/ it has a user guide which is "how to use this" stuff, but then individual classes have documentation like [this one](https://docs.auto-pi-lot.com/en/latest/hardware/gpio.html#autopilot.hardware.gpio.GPIO) which is generated from its [docstring](https://github.com/auto-pi-lot/autopilot/blob/04b5968ba02c8a1413a27eb6a138b6a186b130f1/autopilot/hardware/gpio.py#L135-L165) .

I offered to help out with further questions on docs in case you have never set up a sphinx/readthedocs site before, so we could use this issue to chat if you'd like to pursue that - like i said not all packaged need a docs site, and the README does get the point across for basic usage, but imo it's a really good practice for long-term maintenance to document how the internals of the package work.
