### 2023-10-11 juanmc2005/diart#156 @sneakers-the-rat
I offered to help with how to write tests! So am happy to answer questions/give examples here :)
