### 2023-10-16 3d6330e
bump pyannote to 3.0
