+++ setup.cfg
@@ -31,7 +31,7 @@ install_requires=
     torch>=1.12.1
     torchvision>=0.14.0
     torchaudio>=2.0.2
-    pyannote.audio>=2.1.1
+    pyannote.audio>=3.0.0
     pyannote.core>=4.5
     pyannote.database>=4.1.1
     pyannote.metrics>=3.2
