### 2023-10-18 693c395
add weights normalization, cpu for numpy conversion
