### 2023-10-19 58d86b9
Unify hparam naming. Clean some typing annotations
