### 2023-10-19 22e5633
Remove config.from_dict(). Add --duration argument to CLI
