+++ src/diart/console/benchmark.py
@@ -2,7 +2,10 @@ import argparse
 from pathlib import Path
 
 import pandas as pd
+import torch
+
 from diart import argdoc
+from diart import models as m
 from diart import utils
 from diart.inference import Benchmark, Parallelize
 
@@ -36,6 +39,11 @@ def run():
         "--reference",
         type=Path,
         help="Optional. Directory with RTTM files CONVERSATION.rttm. Names must match audio files",
+    )
+    parser.add_argument(
+        "--duration",
+        type=float,
+        help=f"{argdoc.DURATION}. Defaults to training segmentation duration",
     )
     parser.add_argument(
         "--step", default=0.5, type=float, help=f"{argdoc.STEP}. Defaults to 0.5"
@@ -93,6 +101,14 @@ def run():
     )
     args = parser.parse_args()
 
+    # Resolve device
+    args.device = torch.device("cpu") if args.cpu else None
+
+    # Resolve models
+    hf_token = utils.parse_hf_token_arg(args.hf_token)
+    args.segmentation = m.SegmentationModel.from_pyannote(args.segmentation, hf_token)
+    args.embedding = m.EmbeddingModel.from_pyannote(args.embedding, hf_token)
+
     pipeline_class = utils.get_pipeline_class(args.pipeline)
 
     benchmark = Benchmark(
@@ -104,7 +120,7 @@ def run():
         batch_size=args.batch_size,
     )
 
-    config = pipeline_class.get_config_class().from_dict(vars(args))
+    config = pipeline_class.get_config_class()(**vars(args))
     if args.num_workers > 0:
         benchmark = Parallelize(benchmark, args.num_workers)
 
