### 2023-10-19 344599a
Remove `PipelineConfig.from_dict()` (#189)
* Unify hparam naming. Clean some typing annotations

* Remove config.from_dict(). Add --duration argument to CLI

* Update README.md accordingly
