+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name=diart
-version=0.7.0
+version=0.8.0
 author=Juan Manuel Coria
 description=Streaming speaker diarization in real-time
 long_description=file: README.md
