+++ src/diart/models.py
@@ -3,18 +3,23 @@ from typing import Optional, Text, Union, Callable, Mapping, TYPE_CHECKING
 
 import torch
 import torch.nn as nn
+from requests import HTTPError
 
 try:
     import pyannote.audio.pipelines.utils as pyannote_loader
     from pyannote.audio import Inference, Model
-    from pyannote.audio.pipelines.speaker_verification import WeSpeakerPretrainedSpeakerEmbedding
+    from pyannote.audio.pipelines.speaker_verification import (
+        WeSpeakerPretrainedSpeakerEmbedding,
+    )
 
     _has_pyannote = True
 except ImportError:
     _has_pyannote = False
 
 if TYPE_CHECKING:
-    PipelineInference = Union[WeSpeakerPretrainedSpeakerEmbedding, Model, Text, Mapping]
+    from pyannote.audio.pipelines.speaker_verification import (
+        WeSpeakerPretrainedSpeakerEmbedding,
+    )
 
 
 class PyannoteLoader:
@@ -23,20 +28,14 @@ class PyannoteLoader:
         self.model_info = model_info
         self.hf_token = hf_token
 
-    def __call__(self) -> nn.Module:
-        return pyannote_loader.get_model(self.model_info, self.hf_token)
+    def __call__(self) -> Union[nn.Module, WeSpeakerPretrainedSpeakerEmbedding]:
+        try:
+            return pyannote_loader.get_model(self.model_info)
+        except HTTPError:
+            return WeSpeakerPretrainedSpeakerEmbedding(self.model_info)
 
 
-class PyannoteWeSpeakerSpeakerEmbeddingLoader:
-    def __init__(self, inference_info):
-        super().__init__()
-        self.inference_info = inference_info
-
-    def __call__(self) -> WeSpeakerPretrainedSpeakerEmbedding:
-        return WeSpeakerPretrainedSpeakerEmbedding(self.inference_info)
-
-
-class LazyModel(nn.Module, ABC):
+class LazyModel(ABC):
     def __init__(self, loader: Callable[[], nn.Module]):
         super().__init__()
         self.get_model = loader
@@ -52,35 +51,17 @@ class LazyModel(nn.Module, ABC):
 
     def to(self, *args, **kwargs) -> nn.Module:
         self.load()
-        return super().to(*args, **kwargs)
+        return self.model.to(*args, **kwargs)
 
     def __call__(self, *args, **kwargs):
         self.load()
-        return super().__call__(*args, **kwargs)
+        return self.model(*args, **kwargs)
 
-
-class LazyWeSpeakerSpeakerEmbedding(WeSpeakerPretrainedSpeakerEmbedding, ABC):
-    def __init__(self, loader: Callable[[], WeSpeakerPretrainedSpeakerEmbedding]):
-        self.get_inference = loader
-        self.inference: Optional[WeSpeakerPretrainedSpeakerEmbedding] = None
-        # __init__ at end because in there we call .to() which requires .load()
-        super().__init__()
-
-    def is_in_memory(self) -> bool:
-        """Return whether the model has been loaded into memory"""
-        return self.inference is not None
-
-    def load(self):
-        if not self.is_in_memory():
-            self.inference = self.get_inference()
-
-    def to(self, *args, **kwargs) -> WeSpeakerPretrainedSpeakerEmbedding:
+    def eval(self) -> "LazyModel":
         self.load()
-        return super().to(*args, **kwargs)
-
-    def __call__(self, *args, **kwargs):
-        self.load()
-        return super().__call__(*args, **kwargs)
+        if not isinstance(self.model, WeSpeakerPretrainedSpeakerEmbedding):
+            self.model.eval()
[…]
