### 2023-10-27 6aa756c
Create lint workflow for PRs
