+++ .github/workflows/lint.yml
@@ -1,0 +1,18 @@
+name: Lint
+
+on:
+  pull_request:
+    branches:
+      - master
+      - develop
+
+jobs:
+  lint:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: psf/black@stable
+        with:
+          options: "--check --verbose"
+          src: "./src/diart"
+          version: "23.10.1"
