### 2023-10-27 ccd7d12
remove try catch
