### 2023-10-27 aa2e03b
always normalize
