### 2023-10-27 d0788bb
use PretrainedSpeakerEmbedding in Loader
