### 2023-10-28 714eae3
Fix min-max normalization equation
