### 2023-10-30 d8b2c4f
Change embedding model return type to Callable
Co-authored-by: Simon <80467011+sorgfresser@users.noreply.github.com>
