### 2023-10-30 653984c
remove from active if NaN embeddings
