### 2023-10-30 539e401
Fix wrong typing of model in `LazyModel`
