### 2023-10-31 64056de
Fix wrong path in run-tests.yml
