### 2023-10-31 2eefc81
Add missing arguments in run-tests.yml
