### 2023-10-31 488c7b7
Add initial quick-runs.yml workflow
