+++ .github/workflows/quick-runs.yml
@@ -1,0 +1,43 @@
+# This workflow will install Python dependencies, run tests and lint with a variety of Python versions
+# For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
+
+name: Quick Runs
+
+on:
+  pull_request:
+    branches: [ "main", "develop" ]
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.8"]
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v3
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Download data
+      run: |
+        mkdir audio rttms trash
+        wget --no-verbose --show-progress --continue -O audio/ES2002a.wav http://groups.inf.ed.ac.uk/ami/AMICorpusMirror/amicorpus/ES2002a/audio/ES2002a.Mix-Headset.wav
+        wget --no-verbose --show-progress --continue -O audio/ES2002b.wav http://groups.inf.ed.ac.uk/ami/AMICorpusMirror/amicorpus/ES2002b/audio/ES2002b.Mix-Headset.wav
+        wget --no-verbose --show-progress --continue -O rttms/ES2002a.rttm https://raw.githubusercontent.com/pyannote/AMI-diarization-setup/main/only_words/rttms/train/ES2002a.rttm
+        wget --no-verbose --show-progress --continue -O rttms/ES2002b.rttm https://raw.githubusercontent.com/pyannote/AMI-diarization-setup/main/only_words/rttms/train/ES2002b.rttm
+    - name: Install dependencies
+      run: |
+        sudo apt-get -y install libportaudio2=19.6.0
+        sudo apt-get -y install ffmpeg=4.3.6
+        python -m pip install --upgrade pip
+        pip install .
+    - name: Run stream
+      run: |
+        diart.stream audio/ES2002a.wav --output trash --hf-token ${{ secrets.HUGGINGFACE }} --no-plot
+    - name: Run benchmark
+      run: |
+        diart.benchmark audio --reference rttms --batch-size 4 --hf-token ${{ secrets.HUGGINGFACE }}
