### 2023-10-31 cebc109
Update quick-runs.yml
