+++ .github/workflows/quick-runs.yml
@@ -31,7 +31,7 @@ jobs:
         wget --no-verbose --show-progress --continue -O rttms/ES2002b.rttm https://raw.githubusercontent.com/pyannote/AMI-diarization-setup/main/only_words/rttms/train/ES2002b.rttm
     - name: Install dependencies
       run: |
-        sudo apt-get -y install libportaudio2=19.6.0-1.2
+        sudo apt-get -y install libportaudio2=19.6.0-1.1
         sudo apt-get -y install ffmpeg=4.3.6
         python -m pip install --upgrade pip
         pip install .
