### 2023-10-31 6b50338
Update quick-runs.yml
