+++ .github/workflows/quick-runs.yml
@@ -32,7 +32,7 @@ jobs:
     - name: Install dependencies
       run: |
         sudo apt-get -y install libportaudio2=19.6.0-1.1
-        sudo apt-get -y install ffmpeg=4.3.6
+        sudo apt-get -y install ffmpeg=4.3
         python -m pip install --upgrade pip
         pip install .
     - name: Run stream
