### 2023-10-31 7077d48
Update quick-runs.yml
