### 2023-10-31 5f626ba
Update quick-runs.yml
