### 2023-10-31 32783f0
Update quick-runs.yml
