+++ .github/workflows/quick-runs.yml
@@ -25,8 +25,8 @@ jobs:
     - name: Download data
       run: |
         mkdir audio rttms trash
-        wget --no-verbose --show-progress --continue -O audio/ES2002a.wav http://groups.inf.ed.ac.uk/ami/AMICorpusMirror/amicorpus/ES2002a/audio/ES2002a.Mix-Headset.wav
-        wget --no-verbose --show-progress --continue -O audio/ES2002b.wav http://groups.inf.ed.ac.uk/ami/AMICorpusMirror/amicorpus/ES2002b/audio/ES2002b.Mix-Headset.wav
+        wget --no-verbose --show-progress --continue -O audio/ES2002a_long.wav http://groups.inf.ed.ac.uk/ami/AMICorpusMirror/amicorpus/ES2002a/audio/ES2002a.Mix-Headset.wav
+        wget --no-verbose --show-progress --continue -O audio/ES2002b_long.wav http://groups.inf.ed.ac.uk/ami/AMICorpusMirror/amicorpus/ES2002b/audio/ES2002b.Mix-Headset.wav
         wget --no-verbose --show-progress --continue -O rttms/ES2002a.rttm https://raw.githubusercontent.com/pyannote/AMI-diarization-setup/main/only_words/rttms/train/ES2002a.rttm
         wget --no-verbose --show-progress --continue -O rttms/ES2002b.rttm https://raw.githubusercontent.com/pyannote/AMI-diarization-setup/main/only_words/rttms/train/ES2002b.rttm
     - name: Install dependencies
@@ -38,10 +38,12 @@ jobs:
         pip install .
     - name: Crop audio and rttm
       run: |
-        sox audio/ES2002a.wav audio/ES2002a.wav trim 00:40 01:10
-        sox audio/ES2002b.wav audio/ES2002b.wav trim 00:10 00:40
+        sox audio/ES2002a_long.wav audio/ES2002a.wav trim 00:40 00:30
+        sox audio/ES2002b_long.wav audio/ES2002b.wav trim 00:10 00:30
         head -n 4 rttms/ES2002a.rttm > rttms/ES2002a.rttm
         head -n 7 rttms/ES2002b.rttm > rttms/ES2002b.rttm
+        rm audio/ES2002a_long.wav
+        rm audio/ES2002b_long.wav
     - name: Run stream
       run: |
         diart.stream audio/ES2002a.wav --output trash --no-plot --hf-token ${{ secrets.HUGGINGFACE }}
