### 2023-10-31 389559f
Update quick-runs.yml
