+++ .github/workflows/quick-runs.yml
@@ -27,23 +27,27 @@ jobs:
         mkdir audio rttms trash
         wget --no-verbose --show-progress --continue -O audio/ES2002a_long.wav http://groups.inf.ed.ac.uk/ami/AMICorpusMirror/amicorpus/ES2002a/audio/ES2002a.Mix-Headset.wav
         wget --no-verbose --show-progress --continue -O audio/ES2002b_long.wav http://groups.inf.ed.ac.uk/ami/AMICorpusMirror/amicorpus/ES2002b/audio/ES2002b.Mix-Headset.wav
-        wget --no-verbose --show-progress --continue -O rttms/ES2002a.rttm https://raw.githubusercontent.com/pyannote/AMI-diarization-setup/main/only_words/rttms/train/ES2002a.rttm
-        wget --no-verbose --show-progress --continue -O rttms/ES2002b.rttm https://raw.githubusercontent.com/pyannote/AMI-diarization-setup/main/only_words/rttms/train/ES2002b.rttm
-    - name: Install dependencies
+        wget --no-verbose --show-progress --continue -O rttms/ES2002a_long.rttm https://raw.githubusercontent.com/pyannote/AMI-diarization-setup/main/only_words/rttms/train/ES2002a.rttm
+        wget --no-verbose --show-progress --continue -O rttms/ES2002b_long.rttm https://raw.githubusercontent.com/pyannote/AMI-diarization-setup/main/only_words/rttms/train/ES2002b.rttm
+    - name: Install apt dependencies
       run: |
         sudo add-apt-repository ppa:savoury1/ffmpeg4
         sudo apt-get update
         sudo apt-get -y install ffmpeg libportaudio2=19.6.0-1.1 sox
+    - name: Install pip dependencies
+      run: |
         python -m pip install --upgrade pip
         pip install .
     - name: Crop audio and rttm
       run: |
         sox audio/ES2002a_long.wav audio/ES2002a.wav trim 00:40 00:30
         sox audio/ES2002b_long.wav audio/ES2002b.wav trim 00:10 00:30
-        head -n 4 rttms/ES2002a.rttm > rttms/ES2002a.rttm
-        head -n 7 rttms/ES2002b.rttm > rttms/ES2002b.rttm
+        head -n 4 rttms/ES2002a_long.rttm > rttms/ES2002a.rttm
+        head -n 7 rttms/ES2002b_long.rttm > rttms/ES2002b.rttm
         rm audio/ES2002a_long.wav
         rm audio/ES2002b_long.wav
+        rm rttms/ES2002a_long.rttm
+        rm rttms/ES2002b_long.rttm
     - name: Run stream
       run: |
         diart.stream audio/ES2002a.wav --output trash --no-plot --hf-token ${{ secrets.HUGGINGFACE }}
