### 2023-11-01 18ee8e9
Update quick-runs.yml
