+++ .github/workflows/quick-runs.yml
@@ -14,7 +14,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.8"]
+        python-version: ["3.7", "3.8", "3.10"]
 
     steps:
     - uses: actions/checkout@v3
@@ -54,3 +54,6 @@ jobs:
     - name: Run benchmark
       run: |
         diart.benchmark audio --reference rttms --batch-size 4 --hf-token ${{ secrets.HUGGINGFACE }}
+    - name: Run tuning
+      run: |
+        diart.tune audio --reference rttms --batch-size 4 --num-iter 2 --output trash --hf-token ${{ secrets.HUGGINGFACE }}
