### 2023-11-01 b174792
Update quick-runs.yml
