### 2023-11-01 0279f21
Add workflow for quick Python runs (#194)
