### 2023-11-03 e3c6eef
add docstrings
