### 2023-11-03 f57f9d2
Simplify EmbeddingModel.__call__()
