### 2023-11-03 decd29a
Add ONNX segmentation and embedding models
