+++ src/diart/models.py
@@ -1,9 +1,14 @@
+from __future__ import annotations
+
 from abc import ABC, abstractmethod
-from typing import Optional, Text, Union, Callable
+from pathlib import Path
+from typing import Optional, Text, Union, Callable, List, Tuple
 
 import numpy as np
+import onnxruntime
 import torch
 import torch.nn as nn
+import yaml
 from requests import HTTPError
 
 try:
@@ -58,6 +63,63 @@ class PyannoteLoader:
             )
 
 
+class ONNXLoader:
+    def __init__(self, path: Path, input_names: List[str], output_names: List[str]):
+        super().__init__()
+        self.path = path
+        self.input_names = input_names
+        self.output_names = output_names
+
+    def __call__(self) -> ONNXModel:
+        return ONNXModel(self.path, self.input_names, self.output_names)
+
+
+class ONNXModel:
+    def __init__(self, path: Path, input_names: List[str], output_names: List[str]):
+        super().__init__()
+        self.path = path
+        self.input_names = input_names
+        self.output_names = output_names
+        self.device = torch.device("cpu")
+        self.session = None
+        self.recreate_session()
+
+    def recreate_session(self):
+        if self.device.type == "cuda":
+            providers = ["CUDAExecutionProvider"]
+            provider_options = [{"cudnn_conv_algo_search": "DEFAULT"}]
+        else:
+            providers = ["CPUExecutionProvider"]
+            provider_options = None
+        options = onnxruntime.SessionOptions()
+        options.graph_optimization_level = (
+            onnxruntime.GraphOptimizationLevel.ORT_ENABLE_ALL
+        )
+        self.session = onnxruntime.InferenceSession(
+            self.path,
+            sess_options=options,
+            providers=providers,
+            provider_options=provider_options,
+        )
+
+    def to(self, device: torch.device) -> ONNXModel:
+        if device.type != self.device.type:
+            self.device = device
+            self.recreate_session()
+        return self
+
+
+    def __call__(self, *args) -> Tuple[torch.Tensor, ...]:
+        inputs = {
+            name: arg.cpu().numpy().astype(np.float32)
+            for name, arg in zip(self.input_names, args)
+        }
+        outputs = self.session.run(self.output_names, inputs)
+        return tuple(
+            torch.from_numpy(out).float().to(args[0].device) for out in outputs
+        )
+
+
 class LazyModel(ABC):
     def __init__(self, loader: Callable[[], Callable]):
         super().__init__()
@@ -72,7 +134,7 @@ class LazyModel(ABC):
         if not self.is_in_memory():
             self.model = self.get_model()
 
-    def to(self, device: torch.device) -> "LazyModel":
+    def to(self, device: torch.device) -> LazyModel:
         self.load()
         self.model = self.model.to(device)
         return self
@@ -81,7 +143,7 @@ class LazyModel(ABC):
         self.load()
         return self.model(*args, **kwargs)
 
-    def eval(self) -> "LazyModel":
+    def eval(self) -> LazyModel:
         self.load()
         if isinstance(self.model, nn.Module):
             self.model.eval()
@@ -115,6 +177,17 @@ class SegmentationModel(LazyModel):
         """
         assert _has_pyannote, "No pyannote.audio installation found"
         return PyannoteSegmentationModel(model, use_hf_token)
+
+    @staticmethod
+    def from_onnx(model_path: Union[str, Path]) -> "SegmentationModel":
+        return ONNXSegmentationModel(model_path)
+
+    @staticmethod
[…]
