+++ src/diart/console/stream.py
@@ -103,8 +103,8 @@ def run():
 
     # Resolve models
     hf_token = utils.parse_hf_token_arg(args.hf_token)
-    args.segmentation = m.SegmentationModel.from_pyannote(args.segmentation, hf_token)
-    args.embedding = m.EmbeddingModel.from_pyannote(args.embedding, hf_token)
+    args.segmentation = m.SegmentationModel.from_pretrained(args.segmentation, hf_token)
+    args.embedding = m.EmbeddingModel.from_pretrained(args.embedding, hf_token)
 
     # Resolve pipeline
     pipeline_class = utils.get_pipeline_class(args.pipeline)
