+++ models/segmentation/asru2021.yml
@@ -1,0 +1,2 @@
+sample_rate: 16000
+duration: 5.0
