+++ setup.cfg
@@ -40,6 +40,8 @@ install_requires=
     websocket-server>=0.6.4
     websocket-client>=0.58.0
     rich>=12.5.1
+    pyyaml>=6.0.1
+    onnxruntime>=1.16.1
 
 [options.packages.find]
 where=src

[…]
