### 2023-11-06 e5d31e0
add normalize boolean
