+++ src/diart/argdoc.py
@@ -15,3 +15,4 @@ NUM_WORKERS = "Number of parallel workers"
 OUTPUT = "Directory to store the system's output in RTTM format"
 HF_TOKEN = "Huggingface authentication token for hosted models ('true' | 'false' | <token>). If 'true', it will use the token from huggingface-cli login"
 SAMPLE_RATE = "Sample rate of the audio stream"
+NORMALIZE_EMBEDDING_WEIGHTS = "Normalize embedding weights to be in the range [0, 1]. Useful for WeSpeaker embeddings"
