### 2023-11-06 c708aa0
Update setup.cfg
