### 2023-11-06 #5266 @sneakers-the-rat
I did complete the review!

> My comments here, and boxes left unchecked above are to indicate to the authors where things could be improved and to the editor if they need to uphold the checklist.

Review passes from me, but boxes left unchecked are just to indicate things that are not technically met and was leaving that up to you and the author about what to do

- The repo linked at the top of the issue needs to be changed from https://github.com/juanmc2005/StreamingSpeakerDiarization to https://github.com/juanmc2005/diart
- https://github.com/juanmc2005/diart/issues/176 - Docs are README.md, README.md does not include statement of need, but one is included in the PDF in the repository, which should be moved to the README
- https://github.com/juanmc2005/diart/issues/176 - No API-level docs, though the usage docs in the README are sufficient. I've offered to help with API docs
- https://github.com/juanmc2005/diart/issues/156 - No tests, i have offered to help with tests.
- no contribution guidelines - authors see https://www.pyopensci.org/python-package-guide/documentation/repository-files/contributing-file.html or https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors
