### 2023-11-06 2f4e64f
Minor readability improvements
