### 2023-11-07 #5266 @juanmc2005
Hi @Fei-Tao @sneakers-the-rat, yes I will prioritize those issues.

I'm copying the list here so I can update the progress:
- [ ] The repo linked at the top of the issue needs to be changed from https://github.com/juanmc2005/StreamingSpeakerDiarization to https://github.com/juanmc2005/diart --> **I don't have the rights to do this**
- [ ] Add statement of need to README.md (or API docs)
- [ ] API-level docs, though the usage docs in the README are sufficient
- [ ] Tests
- [ ] Contribution guidelines
