### 2023-11-08 #5266 @sneakers-the-rat
I'm conference prepping atm, but when i return i can help get these last things in place.

est. time, 1h if you've never done it - for docs, you already have docstrings, so to avoid you needing to copy/paste those into README i think the simplest thing would be to do a minimal sphinx boilerplate docs site that takes what you've already written in the README (which is totally sufficient for user docs) and then adds `autodoc` calls to document the classes/functions. see [here](https://docs.readthedocs.io/en/stable/intro/getting-started-with-sphinx.html) for sphinx quickstart, [here](https://docs.readthedocs.io/en/stable/tutorial/index.html) for readthedocs, [here](https://www.pyopensci.org/python-package-guide/documentation/write-user-documentation/get-started.html) for example of literal include README, and [here](https://github.com/readthedocs-examples/example-sphinx-basic) for an example docs repo with api docs.

est. time probably 2h if you've never done it - JOSS doesn't require CI, just existence of tests, so rather than requiring you to write full unit tests for every part of the package, why not just wrap your existing benchmark function in a pytest test? See here where i've made a simple example to test the config - https://github.com/sneakers-the-rat/diart/commit/b1a0ccaa35f8b36aa30f978a4bcb16db69652a42 - so then since you already have your expected outputs in the repo, just run the benchmark and check that the outputs of the benchmark match the expected output. That's not a very sensitive test like you would want out of a test suite, but it will tell you if something somewhere in the whole pipeline is broken.

est. time 15m - contributor docs are an important part of setting the social climate of a package, but at this stage it's not mission critical and you have plenty of time to establish norms as you go. reading and including eg the [contributor covenant](https://www.contributor-covenant.org/) and some simple instructions on how you'd like to receive contributions (eg. search for issues first, open a discussion, raise an issue, make a PR to the `dev` branch, etc.) would totally suffice for the sake of completing the checklist.

So close! Sorry for the measly last things, but these will hopefully help keep the project maintainable in the longterm, and good practice for every package!
