### 2023-11-08 bde46ff
feat: add support for powerset segmentation models
