### 2023-11-08 3ecb58c
feat: add support for powerset segmentation models
