+++ src/diart/models.py
@@ -6,6 +6,7 @@ import torch.nn as nn
 
 try:
     import pyannote.audio.pipelines.utils as pyannote_loader
+    from pyannote.audio.utils.powerset import Powerset
 
     _has_pyannote = True
 except ImportError:
@@ -114,7 +115,15 @@ class PyannoteSegmentationModel(SegmentationModel):
         return self.model.specifications.duration
 
     def forward(self, waveform: torch.Tensor) -> torch.Tensor:
-        return self.model(waveform)
+        predictions = self.model(waveform)
+
+        if (specs := self.model.specifications).powerset:
+            max_speakers_per_frame = specs.powerset_max_classes
+            max_speakers_per_chunk = len(specs.classes)
+            powerset = Powerset(max_speakers_per_chunk, max_speakers_per_frame)
+            predictions = powerset.to_multilabel(predictions, soft=False)
+
+        return predictions
 
 
 class EmbeddingModel(LazyModel):
