### 2023-11-09 0644c70
add abstract __call__
