### 2023-11-09 791d688
move __call__ to parent class
