### 2023-11-09 b9e2c96
wip: trying this PowersetAdapter thing
