### 2023-11-09 72b0e60
wip: trying this PowersetAdapter thing
