### 2023-11-09 da517ac
fix: initialize nn.Module before setting attribute
