### 2023-11-09 11181ea
fix: initialize nn.Module before setting attribute
