### 2023-11-09 bd2313f
Fix unresolved duration and sample rate
