### 2023-11-09 c926c44
Remove model checkpoints
