+++ models/segmentation/asru2021.yml
@@ -1,2 +1,0 @@
-sample_rate: 16000
-duration: 5.0
