+++ requirements.txt
@@ -19,4 +19,4 @@ websocket-server>=0.6.4
 websocket-client>=0.58.0
 rich>=12.5.1
 pyyaml>=6.0.1
-onnxruntime>=1.16.1
+onnxruntime-gpu>=1.16.1
