+++ setup.cfg
@@ -41,7 +41,7 @@ install_requires=
     websocket-client>=0.58.0
     rich>=12.5.1
     pyyaml>=6.0.1
-    onnxruntime>=1.16.1
+    onnxruntime-gpu>=1.16.1
 
 [options.packages.find]
 where=src
