### 2023-11-09 143bcc1
ONNX can have only one output
