### 2023-11-09 472c6cc
Clean up useless embedding model subclasses
