### 2023-11-10 dfd7f33
fix: remove Inference import
