### 2023-11-10 0dd713f
fix: black embedding.py
... though it has nothing to do with this PR...
