### 2023-11-10 d5cbf72
Add support for powerset segmentation models (#198)
* feat: add support for powerset segmentation models

* wip: trying this PowersetAdapter thing

* fix: initialize nn.Module before setting attribute

* Fix unresolved duration and sample rate

* Apply suggestions from code review

* fix: remove Inference import

* fix: black embedding.py

... though it has nothing to do with this PR...

---------

Co-authored-by: Juan Coria <juanmc2005@hotmail.com>
