### 2023-11-10 864f366
Remove duration and sample_rate properties from SegmentationModel. Clean up code
