+++ src/diart/console/serve.py
@@ -36,6 +36,7 @@ def run():
     parser.add_argument(
         "--duration",
         type=float,
+        default=5,
         help=f"{argdoc.DURATION}. Defaults to training segmentation duration",
     )
     parser.add_argument(
