+++ requirements.txt
@@ -18,5 +18,4 @@ optuna>=2.10
 websocket-server>=0.6.4
 websocket-client>=0.58.0
 rich>=12.5.1
-pyyaml>=6.0.1
 onnxruntime-gpu>=1.16.1
