### 2023-11-11 aee7bce
Add ONNX compatibility (#204)
* Add ONNX segmentation and embedding models

* Minor readability improvements

* Replace onnxruntime with onnxruntime-gpu

* ONNX can have only one output

* Clean up useless embedding model subclasses

* Remove duration and sample_rate properties from SegmentationModel. Clean up code

* Update README
