+++ src/diart/models.py
@@ -1,7 +1,11 @@
-from abc import ABC, abstractmethod
-from typing import Optional, Text, Union, Callable
+from __future__ import annotations
+
+from abc import ABC
+from pathlib import Path
+from typing import Optional, Text, Union, Callable, List
 
 import numpy as np
+import onnxruntime
 import torch
 import torch.nn as nn
 from requests import HTTPError
@@ -36,7 +40,7 @@ class PowersetAdapter(nn.Module):
         return self.model.audio
 
     def forward(self, waveform: torch.Tensor) -> torch.Tensor:
-        return self.powerset.to_multilabel(self.model(waveform), soft=False)
+        return self.powerset.to_multilabel(self.model(waveform))
 
 
 class PyannoteLoader:
@@ -58,6 +62,58 @@ class PyannoteLoader:
             )
 
 
+class ONNXLoader:
+    def __init__(self, path: str | Path, input_names: List[str], output_name: str):
+        super().__init__()
+        self.path = Path(path)
+        self.input_names = input_names
+        self.output_name = output_name
+
+    def __call__(self) -> ONNXModel:
+        return ONNXModel(self.path, self.input_names, self.output_name)
+
+
+class ONNXModel:
+    def __init__(self, path: Path, input_names: List[str], output_name: str):
+        super().__init__()
+        self.path = path
+        self.input_names = input_names
+        self.output_name = output_name
+        self.device = torch.device("cpu")
+        self.session = None
+        self.recreate_session()
+
+    @property
+    def execution_provider(self) -> str:
+        device = "CUDA" if self.device.type == "cuda" else "CPU"
+        return f"{device}ExecutionProvider"
+
+    def recreate_session(self):
+        options = onnxruntime.SessionOptions()
+        options.graph_optimization_level = (
+            onnxruntime.GraphOptimizationLevel.ORT_ENABLE_ALL
+        )
+        self.session = onnxruntime.InferenceSession(
+            self.path,
+            sess_options=options,
+            providers=[self.execution_provider],
+        )
+
+    def to(self, device: torch.device) -> ONNXModel:
+        if device.type != self.device.type:
+            self.device = device
+            self.recreate_session()
+        return self
+
+    def __call__(self, *args) -> torch.Tensor:
+        inputs = {
+            name: arg.cpu().numpy().astype(np.float32)
+            for name, arg in zip(self.input_names, args)
+        }
+        output = self.session.run([self.output_name], inputs)[0]
+        return torch.from_numpy(output).float().to(args[0].device)
+
+
 class LazyModel(ABC):
     def __init__(self, loader: Callable[[], Callable]):
         super().__init__()
@@ -72,7 +128,7 @@ class LazyModel(ABC):
         if not self.is_in_memory():
             self.model = self.get_model()
 
-    def to(self, device: torch.device) -> "LazyModel":
+    def to(self, device: torch.device) -> LazyModel:
         self.load()
         self.model = self.model.to(device)
         return self
@@ -81,7 +137,7 @@ class LazyModel(ABC):
         self.load()
         return self.model(*args, **kwargs)
 
-    def eval(self) -> "LazyModel":
+    def eval(self) -> LazyModel:
         self.load()
         if isinstance(self.model, nn.Module):
             self.model.eval()
@@ -114,17 +170,24 @@ class SegmentationModel(LazyModel):
         wrapper: SegmentationModel
         """
         assert _has_pyannote, "No pyannote.audio installation found"
-        return PyannoteSegmentationModel(model, use_hf_token)
+        return SegmentationModel(PyannoteLoader(model, use_hf_token))
 
[…]
