+++ src/diart/blocks/vad.py
@@ -27,11 +27,12 @@ class VoiceActivityDetectionConfig(base.PipelineConfig):
     def __init__(
         self,
         segmentation: m.SegmentationModel | None = None,
-        duration: float | None = None,
+        duration: float = 5,
         step: float = 0.5,
         latency: float | Literal["max", "min"] | None = None,
         tau_active: float = 0.6,
         device: torch.device | None = None,
+        sample_rate: int = 16000,
         **kwargs,
     ):
         # Default segmentation model is pyannote/segmentation
@@ -41,7 +42,7 @@ class VoiceActivityDetectionConfig(base.PipelineConfig):
 
         self._duration = duration
         self._step = step
-        self._sample_rate: int | None = None
+        self._sample_rate = sample_rate
 
         # Latency defaults to the step duration
         self._latency = latency
@@ -57,9 +58,6 @@ class VoiceActivityDetectionConfig(base.PipelineConfig):
 
     @property
     def duration(self) -> float:
-        # Default duration is the one given by the segmentation model
-        if self._duration is None:
-            self._duration = self.segmentation.duration
         return self._duration
 
     @property
@@ -72,9 +70,6 @@ class VoiceActivityDetectionConfig(base.PipelineConfig):
 
     @property
     def sample_rate(self) -> int:
-        # Expected sample rate is given by the segmentation model
-        if self._sample_rate is None:
-            self._sample_rate = self.segmentation.sample_rate
         return self._sample_rate
 
 
