### 2023-11-11 2c52ced
Add link to accept conditions of segmentation 3.0
