### 2023-11-11 8b83ce2
Add table with available models. Add some latencies
