### 2023-11-11 c15e395
hotfix: Catch ModuleNotFoundError when loading a model with pyannote
