### 2023-11-11 6cd0422
Add more info on selecting different models
