### 2023-11-11 4ace494
Add missing info on available models
