### 2023-11-11 dc52922
Add dark themed logo
