+++ README.md
@@ -1,7 +1,7 @@
 <br/>
 
 <p align="center">
-<img width="50%" src="https://raw.githubusercontent.com/juanmc2005/diart/main/logo.jpg" title="Logo" />
+<img width="100%" src="https://raw.githubusercontent.com/juanmc2005/diart/main/logo.jpg" title="diart logo" />
 </p>
 
 <p align="center">
