### 2023-11-13 240b3ac
Add some docstrings
