### 2023-11-13 4d054d2
Ignore private attrs in docs
