### 2023-11-13 b2547d6
Add some docstrings. Effectively ignore __init__
