### 2023-11-13 6041c77
Add documentation page (#209)
* Add initial docs

* Include README in docs page

* Improve README

* Update README

* Add docs requirements.txt

* Add readthedocs config file

* Fix links

* Add some docstrings

* Ignore private attrs in docs

* Add some docstrings. Effectively ignore __init__

* Blacken code

* Blacken code with good version

* Clean up some code

* Fix wrong html title
