### 2023-11-15 fa16c37
Separate huggingface links from model name
