### 2023-11-15 84dc122
Add animated diarization pipeline diagram
