### 2023-11-15 fbfe2a7
Improve pipeline gif
