### 2023-11-16 ac6b4fe
Update snippet gif. Fix torch multiprocessing crash with pyannote 3.1. Other README improvements
