### 2023-11-16 2d0638b
README improvements (#207)
* Add link to accept conditions of segmentation 3.0

* Add table with available models. Add some latencies

* Add more info on selecting different models

* Add missing info on available models

* Improve top menu

* Improve python badge

* Move things around. Simplify code and wording

* Add dark themed logo

* Remove whitespace at the top

* Update README.md

* Rename from_pyannote to from_pretrained in segmentation and embedding blocks

* Separate huggingface links from model name

* Fix reproducibility link

* Add animated diarization pipeline diagram

* Improve 
[…]
