+++ README.md
@@ -1,7 +1,5 @@
-<br/>
-
 <p align="center">
-<img width="50%" src="https://raw.githubusercontent.com/juanmc2005/diart/main/logo.jpg" title="Logo" />
+<img width="100%" src="https://github.com/juanmc2005/diart/blob/main/logo.jpg?raw=true" title="diart logo" />
 </p>
 
 <p align="center">
@@ -11,7 +9,7 @@
 <p align="center">
 <img alt="PyPI Version" src="https://img.shields.io/pypi/v/diart?color=g">
 <img alt="PyPI Downloads" src="https://static.pepy.tech/personalized-badge/diart?period=total&units=international_system&left_color=grey&right_color=brightgreen&left_text=downloads">
-<img alt="Top language" src="https://img.shields.io/github/languages/top/juanmc2005/StreamingSpeakerDiarization?color=g">
+<img alt="Python Versions" src="https://img.shields.io/badge/python-3.8%20%7C%203.9%20%7C%203.10-dark_green">
 <img alt="Code size in bytes" src="https://img.shields.io/github/languages/code-size/juanmc2005/StreamingSpeakerDiarization?color=g">
 <img alt="License" src="https://img.shields.io/github/license/juanmc2005/StreamingSpeakerDiarization?color=g">
 <a href="https://joss.theoj.org/papers/cc9807c6de75ea4c29025c7bd0d31996"><img src="https://joss.theoj.org/papers/cc9807c6de75ea4c29025c7bd0d31996/status.svg"></a>
@@ -27,16 +25,16 @@
       🎙️ Stream audio
     </a>
     <span> | </span>
-    <a href="#-custom-models">
-      🤖 Add your model
-    </a>
-    <span> | </span>
-    <a href="#-tune-hyper-parameters">
-      📈 Tune hyper-parameters
+    <a href="#-models">
+      🧠 Models
     </a>
     <br />
+    <a href="#-tune-hyper-parameters">
+      📈 Tuning
+    </a>
+    <span> | </span>
     <a href="#-build-pipelines">
-      🧠🔗 Build pipelines
+      🧠🔗 Pipelines
     </a>
     <span> | </span>
     <a href="#-websockets">
@@ -45,14 +43,6 @@
     <span> | </span>
     <a href="#-powered-by-research">
       🔬 Research
-    </a>
-    <span> | </span>
-    <a href="#-citation">
-      📗 Citation
-    </a>
-    <span> | </span>
-    <a href="#-reproducibility">
-      👨‍💻 Reproducibility
     </a>
   </h4>
 </div>
@@ -65,38 +55,58 @@
 
 ## ⚡ Quick introduction
 
-Diart is a python framework to build AI-powered real-time audio applications. With diart you can
-create your own AI pipeline, benchmark it, tune its hyper-parameters, and even serve it on the web using websockets.
+Diart is a python framework to build AI-powered real-time audio applications.
+Its key feature is the ability to recognize different speakers in real time with state-of-the-art performance,
+a task commonly known as "speaker diarization".
 
-**We provide pre-trained AI pipelines for:**
+The pipeline `diart.SpeakerDiarization` combines a speaker segmentation and a speaker embedding model
+to power an incremental clustering algorithm that gets more accurate as the conversation progresses:
+
+<p align="center">
+<img width="100%" src="https://github.com/juanmc2005/diart/blob/main/pipeline.gif?raw=true" title="Real-time speaker diarization pipeline" />
+</p>
+
+With diart you can also create your own custom AI pipeline, benchmark it,
+tune its hyper-parameters, and even serve it on the web using websockets.
+
+**We provide pre-trained pipelines for:**
 
 - Speaker Diarization
 - Voice Activity Detection
-- Transcription (coming soon)
[…]
