### 2023-11-16 juanmc2005/diart#176 @juanmc2005
@sneakers-the-rat I updated the README and created a documentation page with all the information you suggested. Everything will be accessible with the next release (v0.9).

Concerning the data structures created, I think it would probably be better to write a tutorial on tuning that could be hosted in readthedocs (see #210).
