### 2023-11-18 5b24a4d
Make ONNX runtime optional
