+++ requirements.txt
@@ -18,4 +18,3 @@ optuna>=2.10
 websocket-server>=0.6.4
 websocket-client>=0.58.0
 rich>=12.5.1
-onnxruntime-gpu>=1.16.1
