### 2023-11-18 8e9f74c
Make ONNX runtime optional (#215)
