+++ setup.cfg
@@ -40,7 +40,6 @@ install_requires=
     websocket-server>=0.6.4
     websocket-client>=0.58.0
     rich>=12.5.1
-    onnxruntime-gpu>=1.16.1
 
 [options.packages.find]
 where=src
