### 2023-11-18 c8e841f
Add link to pyannote 3.1 interpolation issue in README
