### 2023-11-18 7a031d2
Rename from_pyannote to from_pretrained in SpeakerEmbedding
