### 2023-11-18 65a3fd9
Clean unused methods and incorrect arg doc
