### 2023-11-18 9e6c2e9
Clean unused methods and incorrect arg doc
