### 2023-12-11 #5266 @juanmc2005
@sneakers-the-rat concerning the tests for "benchmark" with expected outputs, it's not as straightforward. Expected outputs match the implementation used in the paper, which is not exactly the same as diart (improvements were made since then).
Even then, changes in the future might break that test without it being something bad.

Instead, I would focus on covering (at the very least) all the main functionality in `diart.blocks` (i.e. the `__call__` method of each block class).
