### 2023-12-11 7c650e5
Add testing configuration and diarization tests
