+++ tests/test_diarization.py
@@ -1,0 +1,197 @@
+from __future__ import annotations
+
+import random
+
+import numpy as np
+import pytest
+from pyannote.core import SlidingWindow, SlidingWindowFeature
+
+from diart import SpeakerDiarizationConfig, SpeakerDiarization
+
+
+@pytest.fixture
+def random_diarization_config(
+    segmentation_model, embedding_model
+) -> SpeakerDiarizationConfig:
+    duration = round(random.uniform(1, 10), 1)
+    step = round(random.uniform(0.1, duration), 1)
+    latency = round(random.uniform(step, duration), 1)
+    return SpeakerDiarizationConfig(
+        segmentation=segmentation_model,
+        embedding=embedding_model,
+        duration=duration,
+        step=step,
+        latency=latency,
+    )
+
+
+@pytest.fixture(scope="session")
+def min_latency_config(segmentation_model, embedding_model) -> SpeakerDiarizationConfig:
+    return SpeakerDiarizationConfig(
+        segmentation=segmentation_model,
+        embedding=embedding_model,
+        duration=5,
+        step=0.5,
+        latency="min",
+    )
+
+
+@pytest.fixture(scope="session")
+def max_latency_config(segmentation_model, embedding_model) -> SpeakerDiarizationConfig:
+    return SpeakerDiarizationConfig(
+        segmentation=segmentation_model,
+        embedding=embedding_model,
+        duration=5,
+        step=0.5,
+        latency="max",
+    )
+
+
+def build_waveform_swf(config, start_time: float | None = None) -> SlidingWindowFeature:
+    start_time = round(random.uniform(0, 600), 1) if start_time is None else start_time
+    chunk_size = int(config.duration * config.sample_rate)
+    resolution = config.duration / chunk_size
+    samples = np.random.randn(chunk_size, 1)
+    sliding_window = SlidingWindow(
+        start=start_time, step=resolution, duration=resolution
+    )
+    return SlidingWindowFeature(samples, sliding_window)
+
+
+def test_config(
+    segmentation_model, embedding_model, min_latency_config, max_latency_config
+):
+    duration = round(random.uniform(1, 10), 1)
+    step = round(random.uniform(0.1, duration), 1)
+    latency = round(random.uniform(step, duration), 1)
+    config = SpeakerDiarizationConfig(
+        segmentation=segmentation_model,
+        embedding=embedding_model,
+        duration=duration,
+        step=step,
+        latency=latency,
+    )
+
+    assert config.duration == duration
+    assert config.step == step
+    assert config.latency == latency
+    assert min_latency_config.latency == min_latency_config.step
+    assert max_latency_config.latency == max_latency_config.duration
+
+
+def test_bad_latency(segmentation_model, embedding_model):
+    duration = round(random.uniform(1, 10), 1)
+    step = round(random.uniform(0.5, duration - 0.2), 1)
+    latency_too_low = round(random.uniform(0, step - 0.1), 1)
+    latency_too_high = round(random.uniform(duration + 0.1, 100), 1)
+
+    config1 = SpeakerDiarizationConfig(
+        segmentation=segmentation_model,
+        embedding=embedding_model,
+        duration=duration,
+        step=step,
+        latency=latency_too_low,
+    )
+    config2 = SpeakerDiarizationConfig(
+        segmentation=segmentation_model,
+        embedding=embedding_model,
+        duration=duration,
+        step=step,
+        latency=latency_too_high,
+    )
+
+    with pytest.raises(AssertionError):
+        SpeakerDiarization(config1)
+
[…]
