+++ setup.cfg
@@ -41,6 +41,10 @@ install_requires=
     websocket-client>=0.58.0
     rich>=12.5.1
 
+[options.extras_require]
+tests=
+    pytest>=7.4.0,<8.0.0
+
 [options.packages.find]
 where=src
 
