### 2023-12-13 bfdfdec
Add aggregation tests
