### 2024-01-04 1e1ae2a
speaker embeddings, redis writer
