### 2024-01-04 44a7257
speaker embeddings, redis writer
