### 2024-05-20 #5266 @sneakers-the-rat
Reiterating my willingness to help with tests if the authors lmk what the barriers are here :). Basically I think a simple set of end to end tests that validate the core behavior of the software would be enough for me, not asking the authors to write an entire unit testing suite from the ground up - to me reviewing is both about ensuring some minimum standard of maintainability and functionality for scientific software, but also about teaching and sharing best practices, so am more than happy to help!
