### 2024-05-21 #5266 @juanmc2005
@Fei-Tao @crvernon I was initially thinking of adding unit tests for the entire library, which is why it's been taking me so long.
However, as @sneakers-the-rat pointed out, it seems that what's required is a simple suite of end-to-end tests for the main functionality.

As I mentioned before, I think this end-to-end testing is covered by the `diart.benchmark` CLI, the hyper-parameter values we provide for each dataset, as well as the expected performance that should be observed.
It is not possible to do this automatically for most of the datasets because they're not freely available or too big to be practical. However, I could write a document with step-by-step instructions to test the library on AMI (including how to get the data) and showing the expected performance, or even add an option `--test AMI` to download the data automatically before running.

Once again, the core of this testing is already implemented by `diart.benchmark` and the required configuration is given in the README under the [reproducibility](https://github.com/juanmc2005/diart?tab=readme-ov-file#reproducibility) section, with the exception of downloading the data.

Would that be an acceptable solution?
