### 2024-05-22 #5266 @sneakers-the-rat
I think you should add some sample data that has a known output value and add a simple test with pytest that confirms that the benchmark function reproduces that output :). If you wanted to add unit tests, that only is to your benefit of course, and I would love to see that, but I wouldn't hold the paper up over that. Ideally you would also have CI but the checklist doesn't require that.

like i mentioned here: https://github.com/juanmc2005/diart/issues/175  and in the checklist above, i was able to get output values that were close to those in the paper for one of the datasets, but it required a decent amount of digging around to find the dataset, so for the sake of testing you definitely want to have something small that tests the basic functionality of the package without needing to do a bunch of manual steps to execute the test.

So i would do something like find a 5-10 second sample of particularly challenging audio, run the benchmark over it with all the permutations of the settings you want to test for, confirm that those are correct, and then save those as expected outputs in the tests directory as well.

eg. (some pseudocode bc i'm not going to bother running this right now) assuming the sample is in `tests/data/audio/sample.mp3` and the outputs are in `tests/data/rttm` and this test is `tests/test_benchmark.py`, here is how you would test a range of audio samples and latencies

<summary> expand/collapse test example</summary>

```python
import pytest
from pathlib import Path

from pyannote.database.util import load_rttm

from diart.inference import Benchmark, Parallelize
from diart import SpeakerDiarization, SpeakerDiarizationConfig
from diart.models import SegmentationModel

DATA_DIR = Path(__file__).parent / 'data'
BATCH_SIZE = 32

@pytest.fixture(scope="session")
def model():
    model_name = "pyannote/segmentation@Interspeech2021"
    return SegmentationModel.from_pretrained(model_name)

@pytest.fixture(scope="session")
def make_config(model):
    def _config(latency)
        return SpeakerDiarizationConfig(
            segmentation=model,
            step=0.5,
            latency=latency,
            tau_active=0.555,
            rho_update=0.422,
            delta_new=1.517
        )
    return _config

@pytest.mark.parametrize('source_file', (DATA_DIR / 'audio').glob('*.mp3'))
@pytest.mark.parametrize('latency', [0.5, 1, 2, 3, 4, 5])
def test_benchmark(make_config, source_file, latency):
    config = make_config(latency)
    pipeline = SpeakerDiarization(config)

    padding = pipeline.config.get_file_padding(filepath)
    source = src.FileAudioSource(
        source_file,
        pipeline.config.sample_rate,
        padding,
        pipeline.config.step,
    )

    pipeline.set_timestamp_shift(-padding[0])
    inference = StreamingInference(
        pipeline,
        source,
        BATCH_SIZE,
        do_profile=False,
        do_plot=False,
        show_progress=False
    )

    pred = inference()

    expected_file = (DATA_DIR / 'rttm' / str(latency) / source_file.name).with_suffix('.rttm')
    expected = load_rttm(expected_file)

    assert pred == expected

```
