### 2024-05-22 #5266 @juanmc2005
@sneakers-the-rat thanks for your answer! I can add a test case like this. My only concern is that I don't necessarily want to check for "expected" outputs because what's expected is unclear. There are no "correct" outputs in practice, only if they are perfect like a human annotation would be, and no diarization system can produce human-level outputs yet.

In this context, it's unclear to me what the purpose of the test would be. As soon as an improvement is made on the core algorithm (e.g. clustering) to make it more accurate, the test will become useless because the outputs might be better this time, or they might be worse for the test file but overall better for a benchmark dataset (that is what users will actually care about).
This is actually what happened already with the expected outputs from the paper that I provided at the very beginning, which were improved upon later.
The end result would be that each time I upgrade the library I will also have to update the expected outputs, so it feels like I'm writing a test just for it to pass.

I understand it could be useful to verify that small changes outside the core features don't break a specific version. But a more robust test is to actually run a benchmark on one of the freely available target datasets like AMI, because the outputs might be the same for one file but might break in others.

In any case, I'll go ahead and work on this end-to-end test next weekend. I'll keep you updated about the progress.
