### 2024-05-22 #5266 @juanmc2005
@sneakers-the-rat concerning the CI, I already included github actions for linting and quick testing that the main operations run on both python 3.8 and 3.10 on every PR.
There's also a github action for automatic deployment to pypi when creating a release.
