### 2024-05-22 #5266 @sneakers-the-rat
This is totally fair and a very good point!

I feel multiple ways about this: on the one hand, I dont want to police what "counts" as a test, and the JOSS reviewer guidelines dont describe any standard for coverage, robustness, or completeness of tests (they probably should, but they dont). I also dont want to demand you write a whole ton of code for the sake of this review which has already gone on long enough. On the other hand, I do feel like the purpose of review is to help the software reach some minimal state of maintainability and reliability and that includes testing for core functionality.

It is very true that "snapshot" based end-to-end tests are some of the weakest tests you can write, and that they are mostly useful when accompanied by unit and other more specific tests. They are not entirely pointless, though, as they do signal that something about the program has changed - eg. If your upstream model changes in such a way that it breaks (instead of getting better), you would want to know that even if nothing changed in your code. So while the snapshot style tests are not ideal, they are relatively near at hand without asking you to write a ton of unit tests while still covering core functionality.

I did see the CI action that runs the benchmark, but testing for execution is a few steps weaker still (though still not nothing!) s.t. I dont think they help the package be a maintainable and reliable part of open source tooling - eg. If you accept a large PR and it accidentally starts adding 10 seconds to all timestamps, those tests would still pass even if it breaks the tool.

My example was just a quick illustration and I take your point about the moving target - the key point is having known input and output. So an ideal case would be one where you definitely know the correct times (and the audio signal is still naturalistic, so its not "too easy" bc youre right, you do want reasonable coverage of the input space bc some samples may fail while others still work), and then you would do something like set a threshold like "do the estimated times fall within some threshold window that I would still call correct." If the model gets more accurate, great, that should just push the values closer to the center of that expected window, but you wouldnt have to fiddle with the expected values for some trivial ms change.

Re: the form of the tests, I do think that using a tool like pypi is preferable vs. A simple CLI test because it gives you space to grow for future testing, makes for a better developer experience for potential contributors, and allows you to do more complex tests eg. On the product of conditions audio samples x configuration options as I demo in above comment. The CI action would then just be to run `pytest`. If you prefer to just have tests just as a CLI call I definitely wouldnt force you to use pytest, so consider that a strong but friendly recommendation to bring the package in line with best practices ♥

To avoid cluttering the review issue, if theres more discussion to be had lets move this into an issue on diart. Like I said, good and fair points! Not trying to shut down conversation just keeping things tidy given github's lack of threading on issues
