### 2024-05-25 b819574
Fix rounding error in min latency unit test
