### 2024-05-25 ff3cec7
Improve CI workflows and add pytest. Fix matplotlib colormap error
