+++ setup.cfg
@@ -21,7 +21,7 @@ package_dir=
 packages=find:
 install_requires=
     numpy>=1.20.2
-    matplotlib>=3.3.3
+    matplotlib>=3.3.3,<3.6.0
     rx>=3.2.0
     scipy>=1.6.0
     sounddevice>=0.4.2
