### 2024-05-25 0f41bcb
Install missing dependencies in CI
