+++ .github/workflows/pytest.yml
@@ -28,7 +28,7 @@ jobs:
     - name: Install pip dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install .
+        pip install .[tests]
 
     - name: Run tests
       run: |
