### 2024-05-25 #5266 @juanmc2005
@sneakers-the-rat I just opened the following PR with some unit tests and the end-to-end test we discussed about: https://github.com/juanmc2005/diart/pull/237

I also added a GitHub action to automatically run pytest on all PRs to `main` and `develop`
