### 2024-05-25 09b85bb
Add onnxruntime as a test dependency
