+++ setup.cfg
@@ -44,6 +44,7 @@ install_requires=
 [options.extras_require]
 tests=
     pytest>=7.4.0,<8.0.0
+    onnxruntime==1.18.0
 
 [options.packages.find]
 where=src
