+++ tests/test_diarization.py
@@ -1,0 +1,204 @@
+from __future__ import annotations
+
+import random
+
+import pytest
+
+from diart import SpeakerDiarizationConfig, SpeakerDiarization
+from utils import build_waveform_swf
+
+
+@pytest.fixture
+def random_diarization_config(
+    segmentation_model, embedding_model
+) -> SpeakerDiarizationConfig:
+    duration = round(random.uniform(1, 10), 1)
+    step = round(random.uniform(0.1, duration), 1)
+    latency = round(random.uniform(step, duration), 1)
+    return SpeakerDiarizationConfig(
+        segmentation=segmentation_model,
+        embedding=embedding_model,
+        duration=duration,
+        step=step,
+        latency=latency,
+    )
+
+
+@pytest.fixture(scope="session")
+def min_latency_config(segmentation_model, embedding_model) -> SpeakerDiarizationConfig:
+    return SpeakerDiarizationConfig(
+        segmentation=segmentation_model,
+        embedding=embedding_model,
+        duration=5,
+        step=0.5,
+        latency="min",
+    )
+
+
+@pytest.fixture(scope="session")
+def max_latency_config(segmentation_model, embedding_model) -> SpeakerDiarizationConfig:
+    return SpeakerDiarizationConfig(
+        segmentation=segmentation_model,
+        embedding=embedding_model,
+        duration=5,
+        step=0.5,
+        latency="max",
+    )
+
+
+def test_config(
+    segmentation_model, embedding_model, min_latency_config, max_latency_config
+):
+    duration = round(random.uniform(1, 10), 1)
+    step = round(random.uniform(0.1, duration), 1)
+    latency = round(random.uniform(step, duration), 1)
+    config = SpeakerDiarizationConfig(
+        segmentation=segmentation_model,
+        embedding=embedding_model,
+        duration=duration,
+        step=step,
+        latency=latency,
+    )
+
+    assert config.duration == duration
+    assert config.step == step
+    assert config.latency == latency
+    assert min_latency_config.latency == min_latency_config.step
+    assert max_latency_config.latency == max_latency_config.duration
+
+
+def test_bad_latency(segmentation_model, embedding_model):
+    duration = round(random.uniform(1, 10), 1)
+    step = round(random.uniform(0.5, duration - 0.2), 1)
+    latency_too_low = round(random.uniform(0, step - 0.1), 1)
+    latency_too_high = round(random.uniform(duration + 0.1, 100), 1)
+
+    config1 = SpeakerDiarizationConfig(
+        segmentation=segmentation_model,
+        embedding=embedding_model,
+        duration=duration,
+        step=step,
+        latency=latency_too_low,
+    )
+    config2 = SpeakerDiarizationConfig(
+        segmentation=segmentation_model,
+        embedding=embedding_model,
+        duration=duration,
+        step=step,
+        latency=latency_too_high,
+    )
+
+    with pytest.raises(AssertionError):
+        SpeakerDiarization(config1)
+
+    with pytest.raises(AssertionError):
+        SpeakerDiarization(config2)
+
+
+def test_pipeline_build(random_diarization_config):
+    pipeline = SpeakerDiarization(random_diarization_config)
+
+    assert pipeline.get_config_class() == SpeakerDiarizationConfig
+
+    hparams = pipeline.hyper_parameters()
+    hp_names = [hp.name for hp in hparams]
+    assert len(set(hp_names)) == 3
+
+    for hparam in hparams:
+        assert hparam.low == 0
+        if hparam.name in ["tau_active", "rho_update"]:
+            assert hparam.high == 1
[…]
