### 2024-05-25 juanmc2005/diart#156 @juanmc2005
Test PR #237 has been merged and will be included in the next release (v0.10)
