### 2024-05-30 9d8e20e
Updated readme embed-extraction pipeline
