### 2024-06-23 #5266 @juanmc2005
Hi @Fei-Tao, I've just published the latest release (0.9.1) to Zenodo. Here's the DOI: [10.5281/zenodo.12510278](https://zenodo.org/records/12510278).

I already went through the checklist and everything looks good from my side. Thanks!
