### 2024-06-28 14f96a7
Limit numpy version to < 2.0.0
