+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy>=1.20.2
+numpy>=1.20.2,<2.0.0
 matplotlib>=3.3.3,<3.6.0
 rx>=3.2.0
 scipy>=1.6.0
