+++ setup.cfg
@@ -20,7 +20,7 @@ package_dir=
     =src
 packages=find:
 install_requires=
-    numpy>=1.20.2
+    numpy>=1.20.2,<2.0.0
     matplotlib>=3.3.3,<3.6.0
     rx>=3.2.0
     scipy>=1.6.0
