### 2024-07-01 #5266 @crvernon
:wave: @juanmc2005 - I just have the following few things I need you to address before I move to accept this publication:

- [ ] The version specified in your Zenodo archive (v1) does not match the version decalared in the review thread (v0.9.1) see https://zenodo.org/records/12510278.  Please edit this to match.
- [ ] The end of the "Summary" section in the paper is abrupt and never discusses the introduction of your proposed solution.  I would state something like the following as a last sentence in the "Summary" section.  "We introduce `Diart` as Python package to efficiently conduct real-time speaker diarization."

Thanks!
